Data partitioning for training and testing models

This function provides different partition methods based in folds (kfold, rep_kfold, and loocv), bootstrap (boot), and spatially structured (BANDS and BLOCK).

data_part(data, p_a, bg_data = NULL, bg_a = NULL, method = NULL)

Arguments

data
data.frame. Database with presences, presence-absence, or pseudo-absence, records for a given species

p_a
character. Column name of "data" with presences, presence-absence, or pseudo-absence. Presences must be represented by 1 and absences by 0

bg_data
data.frame. Data frames with background points.

bg_a
character. Column name of "bg_data" with absences. Background must be repented by 0

method
character. Vector with partition method to be used:

kfold: Random partition in k-fold cross-validation. Usage part=c(method= 'kfold', folds='5'). 'folds' refers to the number of folds for data partitioning, it assumes value >=1. Usage method = c(method = "kfold", folds = 10).

rep_kfold: Random partition in repeated k-fold cross-validation. Usage method = c(method = "rep_kfold", folds = 10, replicates=10). 'folds' refers to the number of folds for data partitioning, it assumes value >=1. 'replicate' refers to the number of replicates, it assumes a value >=1.

loocv: Leave-one-out cross-validation (a.k.a. Jackknife) WRITE METHOD HERE . Usage method = c(method = "loocv").

boot: Random bootstrap partition. Usage part=c(method='boot', replicates='2', proportion='0.7'). 'replicate' refers to the number of replicates, it assumes a value >=1. 'proportion' refers to the proportion of occurrences used for model fitting, it assumes a value >0 and <=1. In this example proportion='0.7' mean that 70

Value

This function will return the same data.frame used in the arguments 'data' and 'gb_data' with the additional column or columns staring with the names .part with partition group. In the case of boot partition data will be stored with 'train' and 'test' words, for all other methods partition groups will be stored as numeric. In case of use 'data' and 'bg_data' will returned a list with partition data for each database with the names data and bg_data.

See also

Examples

if (FALSE) {
require(devtools)
data("abies_db")
abies_db$partition <- NULL
abies_db <- tibble(abies_db)

# K-fold method
abies_db2 <- data_part(
  data = abies_db,
  p_a = "pr_ab",
  bg_data = NULL,
  bg_a = NULL,
  method = c(method = "kfold", folds = 10)
)
abies_db2

# Repeated K-fold method
abies_db2 <- data_part(
  data = abies_db,
  p_a = "pr_ab",
  bg_data = NULL,
  bg_p_a = NULL,
  method = c(method = "rep_kfold", folds = 10, replicates = 10)
)
abies_db2

# Leave-one-out cross-validation (loocv) method
abies_db2 <- data_part(
  data = abies_db,
  p_a = "pr_ab",
  bg_data = NULL,
  bg_a = NULL,
  method = c(method = "loocv")
)
abies_db2

# Bootstrap method
abies_db2 <- data_part(
  data = abies_db,
  p_a = "pr_ab",
  bg_data = NULL,
  bg_a = NULL,
  method = c(method = "boot", replicates = 50, proportion = 0.7)
)
abies_db2$.part1 %>% table() # Note that for this method .partX columns have train and test words.
}