Calculate different model performance metrics

This function calculates threshold dependent and independent performance metric. It calculates TPR (True Positive Rate, also called sensitivity), TNR (True Negative Rate, , also called specificity), SORENSEN, JACCARD, FPB, OR (Omission Rate), TSS (True Skill Statistic) KAPPA, AUC (Area Under Curve), and BOYCE.

enm_eval(p, a, bg = NULL, thr = NULL)

Arguments

p
numeric. Predicted suitability for presences

a
numeric. Predicted suitability for presences absences

bg
numeric. Predicted suitability for background points. It is used for BOYCE metric. It bg is set as null BOYCE metric will be calculated with presences and absences suitability values

thr
character. Threshold used to get binary suitability values (i.e. 0,1). It is useful for threshold-dependent performance metrics. It is possible to use more than one threshold type. It is necessary to provide a vector for this argument. The next threshold area available:

LPT: The highest threshold at which there is no omission. Usage thr=c(type='LPT').

EQUAL_SENS_SPEC: Threshold at which the sum of the sensitivity and specificity is the highest.

MAX_TSS: Threshold at which the sensitivity and specificity are equal. Usage thr=c(type='MAX_TSS').

MAX_KAPPA: The threshold at which kappa is the highest ("max kappa"). Usage thr=c(type='MAX_KAPPA').

MAX_JACCARD: The threshold at which Jaccard is the highest. Usage thr=c(type='MAX_JACCARD').

MAX_SORENSEN: The threshold at which Sorensen is highest. Usage thr=c(type='MAX_SORENSEN').

MAX_FPB: The threshold at which FPB is highest. Usage thr=c(type='MAX_FPB').

SENSITIVITY: A threshold value specified by user. Usage thr=c(type='SENSITIVITY', sens='0.6'). 'sens' refers to models will be binarized using this suitability value.

In the case of use more than one threshold type it is necessary concatenate the names of threshold types, e.g., thr=c(type=c('LPT', 'MAX_TSS', 'MAX_JACCARD')). When SENSITIVITY threshold is used in combination with other it is necessary specify the desired sensitivity value, e.g. thr=c(type=c('LPT', 'MAX_TSS', 'SENSITIVITY'), sens='0.8')

Value

a list with the next tibble

"performance". A tibble object with the performance metric for 90 threshold values >0 and <1. Performance.

"threshold". A tibble with values for all the threshold available in the package.

"threshold_table". A tibble with the threshold values and the performance metric values for each threshold.

" selected_threshold". It is similar to "threshold_table" with difference that it contains the thresholds and the performance metric values for the selected threshold. A tibble with threshold values and the performance metric values for each threshold.

when thr argument is NULL function will return a list with "performance", "threshold", and "threshold_table" tibbles, when thr argument is used with one or more threshold function returns "selected_threshold" and "threshold_table" tibbles

Examples

if (FALSE) {
require(dismo)
require(dplyr)

set.seed(0)
p <- rnorm(50, mean=0.7, sd=0.3) %>% abs()
p[p>1] <- 1
p[p<0] <- 0

set.seed(0)
a <- rnorm(50, mean=0.4, sd=0.4) %>% abs()
a[a>1] <- 1
a[a<0] <- 0

set.seed(0)
backg <- rnorm(1000, mean=0.4, sd=0.4) %>% abs()
backg[backg>1] <- 1
backg[backg<0] <- 0

# Use function without threshold specification 
e <- enm_eval(p, a)
e$performance
e$threshold
e$threshold_table

enm_eval(p, a, thr=c(type=c('MAX_KAPPA')))
enm_eval(p, a, thr=c(type=c('LPT', 'MAX_TSS', 'MAX_JACCARD')))
enm_eval(p, a, thr=c(type=c('LPT', 'MAX_TSS', 'SENSITIVITY'))) # wrong way to SENSITIVITY threshold
enm_eval(p, a, thr=c(type=c('LPT', 'MAX_TSS', 'SENSITIVITY'), sens='0.8')) # correct way to use SENSITIVITY threshold

# Use of bg argument (it will only be used for calculating BOYCE index)
enm_eval(p, a, thr=c(type=c('MAX_TSS')))[[1]]
enm_eval(p, a, thr=c(type=c('MAX_TSS')), bg=backg)[[1]]
# I the case it is needed use background for calculate all other metric background values can be used in "a" argument 
enm_eval(p, backg, thr=c(type=c('MAX_TSS')))[[1]]
}