Raster interpolation between two time periods

This function calculates suitability values between two time periods with simple interpolation using two raster objects with suitability values.

inter(r1, r2, y1, y2, rastername = NULL, dir_save = NULL)

Arguments

r1
SpatRaster. Raster object for the initial year

r2
SpatRaster. Raster object for the final year

y1
numeric. Initial year

y2
numeric. Final year

rastername
character. Word used as prefix in raster file name. Default NULL

dir_save
character. Directory path and name of the folder in which you want to save the raster files. If NULL, function will return a SpatRaster object, on the contrary, it will save raster in a given directory. Default NULL

Value

This function returns a SpatRaster if dir_save is used as NULL, If dirsave is used, function will return anythink

Examples

if (FALSE) {
require(terra)
require(dplyr)

f <- system.file("external/suit_time_step.tif", package = "flexsdm")
abma <- terra::rast(f)
plot(abma)

int <- inter(
  r1 = abma[[1]],
  r2 = abma[[2]],
  y1 = 2010,
  y2 = 2020,
  rastername = "Abies",
  dir_save = NULL
)

int
}