A data set containing presences of palms species from Southern Brazil

This data set contains presences of 11 palm species from Southern Brazil sourced by Calambás-Trochez et al. (2021).

Usage

palms

Format

A tibble with 327 rows and 3 variables:

species

species names

x

longitude of species occurrences

y

latitude of species occurrences

...

References

Calambás-Trochez, L.F., Velazco, S.J.E., Hoffmann, P.M., Brum, F.T., Carlucci, M.B., 2021. Climate and land-use changes coupled with low coverage of protected areas threaten palm species in South Brazilian grasslands. Perspectives in Ecology and Conservation 9. https://doi.org/10.1016/j.pecon.2021.03.010

Examples

if (FALSE) { # \dontrun{
require(dplyr)
data("palms")
} # }