Data partitioning for training and testing models

This function provides different conventional partition methods based in folds (kfold, rep_kfold, and loocv), and bootstrap (boot)

part_classical(data, pr_ab, method = NULL)

Arguments

data
data.frame. Database with presences, presence-absence, or pseudo-absence, records for a given species

pr_ab
character. Column name of "data" with presences, presence-absence, or pseudo-absence. Presences must be represented by 1 and absences by 0

method
character. Vector with partition method to be used:

kfold: Random partition in k-fold cross-validation. Usage part=c(method= 'kfold', folds='5'). 'folds' refers to the number of folds for data partitioning, it assumes value >=1. Usage method = c(method = "kfold", folds = 10).

rep_kfold: Random partition in repeated k-fold cross-validation. Usage method = c(method = "rep_kfold", folds = 10, replicates=10). 'folds' refers to the number of folds for data partitioning, it assumes value >=1. 'replicate' refers to the number of replicates, it assumes a value >=1.

loocv: Leave-one-out cross-validation (a.k.a. Jackknife) WRITE METHOD HERE . Usage method = c(method = "loocv").

boot: Random bootstrap partition. Usage part=c(method='boot', replicates='2', proportion='0.7'). 'replicate' refers to the number of replicates, it assumes a value >=1. 'proportion' refers to the proportion of occurrences used for model fitting, it assumes a value >0 and <=1. In this example proportion='0.7' mean that 70

Value

A tibble object with information used in 'data' argument and a additional columns .part with partition group. In the case of boot partition partition data will be stored with 'train' and 'test' words, for all other methods partition groups will be stored as numeric.

See also

Examples

# \dontrun{
data("abies_db")
abies_db$partition <- NULL
abies_db <- tibble(abies_db)

# K-fold method
abies_db2 <- part(
  data = abies_db,
  pr_ab = "pr_ab",
  method = c(method = "kfold", folds = 10)
)
abies_db2

#> # A tibble: 1,309 x 12
#>    pr_ab   aet   cwd   tmin ppt_djf ppt_jja    pH    awc depth percent_clay
#>    <int> <dbl> <dbl>  <dbl>   <dbl>   <dbl> <dbl>  <dbl> <dbl>        <dbl>
#>  1     0  187.  672. -0.320    75.2    8.99  6.70 0.120   74.7        21.9 
#>  2     0  412.  755.  9.38    172.    10.3   6.10 0.0849  42.9        27.4 
#>  3     0  366.  814.  9.86    129.     7.57  6.40 0.0900  46          32.7 
#>  4     0  351.  817.  9.42    129.     7.70  6.00 0.0980  35.6        24.6 
#>  5     0  370.  873.  9.14    135.     5.02  7.09 0.130   20.2        33.8 
#>  6     0  400.  789. 10.0     184.    10.0   6.10 0.0600  28          18.5 
#>  7     0  423.  735.  9.68    210.    11.2   2.59 0.0382  11.9         7.86
#>  8     0  362.  360. -1.38    155.    20.6   6.40 0.0400 201           2.20
#>  9     0  412.  208. -2.09    455.    27.1   5.83 0.0476 174.          1.59
#> 10     1  500.  173. -1.74    389.    25.6   6.14 0.0581  77.7         3.07
#> # ... with 1,299 more rows, and 2 more variables: landform <fct>, .part <int>


# Repeated K-fold method
abies_db2 <- part(
  data = abies_db,
  pr_ab = "pr_ab",
  method = c(method = "rep_kfold", folds = 10, replicates = 10)
)
abies_db2

#> # A tibble: 1,309 x 21
#>    pr_ab   aet   cwd   tmin ppt_djf ppt_jja    pH    awc depth percent_clay
#>    <int> <dbl> <dbl>  <dbl>   <dbl>   <dbl> <dbl>  <dbl> <dbl>        <dbl>
#>  1     0  187.  672. -0.320    75.2    8.99  6.70 0.120   74.7        21.9 
#>  2     0  412.  755.  9.38    172.    10.3   6.10 0.0849  42.9        27.4 
#>  3     0  366.  814.  9.86    129.     7.57  6.40 0.0900  46          32.7 
#>  4     0  351.  817.  9.42    129.     7.70  6.00 0.0980  35.6        24.6 
#>  5     0  370.  873.  9.14    135.     5.02  7.09 0.130   20.2        33.8 
#>  6     0  400.  789. 10.0     184.    10.0   6.10 0.0600  28          18.5 
#>  7     0  423.  735.  9.68    210.    11.2   2.59 0.0382  11.9         7.86
#>  8     0  362.  360. -1.38    155.    20.6   6.40 0.0400 201           2.20
#>  9     0  412.  208. -2.09    455.    27.1   5.83 0.0476 174.          1.59
#> 10     1  500.  173. -1.74    389.    25.6   6.14 0.0581  77.7         3.07
#> # ... with 1,299 more rows, and 11 more variables: landform <fct>,
#> #   .part1 <int>, .part2 <int>, .part3 <int>, .part4 <int>, .part5 <int>,
#> #   .part6 <int>, .part7 <int>, .part8 <int>, .part9 <int>, .part10 <int>


# Leave-one-out cross-validation (loocv) method
abies_db2 <- part(
  data = abies_db,
  pr_ab = "pr_ab",
  method = c(method = "loocv")
)
abies_db2

#> # A tibble: 1,309 x 12
#>    pr_ab   aet   cwd   tmin ppt_djf ppt_jja    pH    awc depth percent_clay
#>    <int> <dbl> <dbl>  <dbl>   <dbl>   <dbl> <dbl>  <dbl> <dbl>        <dbl>
#>  1     0  187.  672. -0.320    75.2    8.99  6.70 0.120   74.7        21.9 
#>  2     0  412.  755.  9.38    172.    10.3   6.10 0.0849  42.9        27.4 
#>  3     0  366.  814.  9.86    129.     7.57  6.40 0.0900  46          32.7 
#>  4     0  351.  817.  9.42    129.     7.70  6.00 0.0980  35.6        24.6 
#>  5     0  370.  873.  9.14    135.     5.02  7.09 0.130   20.2        33.8 
#>  6     0  400.  789. 10.0     184.    10.0   6.10 0.0600  28          18.5 
#>  7     0  423.  735.  9.68    210.    11.2   2.59 0.0382  11.9         7.86
#>  8     0  362.  360. -1.38    155.    20.6   6.40 0.0400 201           2.20
#>  9     0  412.  208. -2.09    455.    27.1   5.83 0.0476 174.          1.59
#> 10     1  500.  173. -1.74    389.    25.6   6.14 0.0581  77.7         3.07
#> # ... with 1,299 more rows, and 2 more variables: landform <fct>, .part <int>


# Bootstrap method
abies_db2 <- part(
  data = abies_db,
  pr_ab = "pr_ab",
  method = c(method = "boot", replicates = 50, proportion = 0.7)
)
abies_db2

#> # A tibble: 1,309 x 61
#>    pr_ab   aet   cwd   tmin ppt_djf ppt_jja    pH    awc depth percent_clay
#>    <int> <dbl> <dbl>  <dbl>   <dbl>   <dbl> <dbl>  <dbl> <dbl>        <dbl>
#>  1     0  187.  672. -0.320    75.2    8.99  6.70 0.120   74.7        21.9 
#>  2     0  412.  755.  9.38    172.    10.3   6.10 0.0849  42.9        27.4 
#>  3     0  366.  814.  9.86    129.     7.57  6.40 0.0900  46          32.7 
#>  4     0  351.  817.  9.42    129.     7.70  6.00 0.0980  35.6        24.6 
#>  5     0  370.  873.  9.14    135.     5.02  7.09 0.130   20.2        33.8 
#>  6     0  400.  789. 10.0     184.    10.0   6.10 0.0600  28          18.5 
#>  7     0  423.  735.  9.68    210.    11.2   2.59 0.0382  11.9         7.86
#>  8     0  362.  360. -1.38    155.    20.6   6.40 0.0400 201           2.20
#>  9     0  412.  208. -2.09    455.    27.1   5.83 0.0476 174.          1.59
#> 10     1  500.  173. -1.74    389.    25.6   6.14 0.0581  77.7         3.07
#> # ... with 1,299 more rows, and 51 more variables: landform <fct>,
#> #   .part1 <chr>, .part2 <chr>, .part3 <chr>, .part4 <chr>, .part5 <chr>,
#> #   .part6 <chr>, .part7 <chr>, .part8 <chr>, .part9 <chr>, .part10 <chr>,
#> #   .part11 <chr>, .part12 <chr>, .part13 <chr>, .part14 <chr>, .part15 <chr>,
#> #   .part16 <chr>, .part17 <chr>, .part18 <chr>, .part19 <chr>, .part20 <chr>,
#> #   .part21 <chr>, .part22 <chr>, .part23 <chr>, .part24 <chr>, .part25 <chr>,
#> #   .part26 <chr>, .part27 <chr>, .part28 <chr>, .part29 <chr>, .part30 <chr>,
#> #   .part31 <chr>, .part32 <chr>, .part33 <chr>, .part34 <chr>, .part35 <chr>,
#> #   .part36 <chr>, .part37 <chr>, .part38 <chr>, .part39 <chr>, .part40 <chr>,
#> #   .part41 <chr>, .part42 <chr>, .part43 <chr>, .part44 <chr>, .part45 <chr>,
#> #   .part46 <chr>, .part47 <chr>, .part48 <chr>, .part49 <chr>, .part50 <chr>

abies_db2$.part1 %>% table() # Note that for this method .partX columns have train and test words.

#> .
#>       test      train train-test 
#>        102        625        290 

# }