Spatial block cross validation

This function explores spatial blocks with different cell sizes and returns the best one suited for a given presence or presence-absences database. The selection of the best grid size is performed automatically considering spatial autocorrelation, environmental similarity, and the number of presence and absence records in each partition.

part_spat(
  env_layer,
  data,
  x,
  y,
  pr_ab,
  n_part = 3,
  min_res_mult = 3,
  max_res_mult = 200,
  num_grids = 30,
  prop = 0.5
)

Arguments

env_layer
SpatRaster. Raster with environmental variable. This will be used to evaluate spatial autocorrelation and environmental similarity between training and testing partition. Because this function calculate dissimilarity based on euclidean distances, it can only handle continuous layers, do not use categorical layers as inputs

data
data.frame. Data.frame or tibble object with presences (or presence-absence, or presences-pseudo-absence) records, and coordinates

x
character. Column name with longitude data

y
character. Column name with latitude data

pr_ab
character. Column with presences, presence-absence, or pseudo-absence. Presences must be represented by 1 and absences by 0

n_part
integer. Number of partition. Default 2, values other than 2 has not yet been implemented.

min_res_mult
integer. Minimum value used for multiplying raster resolution and define the finest resolution to be tested, default 3.

max_res_mult
integer. Maximum value used for multiplying raster resolution and define the coarsest resolution to be tested, default 200.

num_grids
integer. Number of grid to be tested between min_res_mult X (raster resolution) and max_res_mult X (raster resolution), default 30

prop
numeric. Proportion of point used for testing autocorrelation between groups (values > 0 and <=1). The smaller this number is, the faster the function will work. Default 0.5

Value

A list with:

part: A tibble object with information used in 'data' arguments and a additional column .part with partition group.

best_part_info: A tibble with information of the bets partition. It contains the number of the best partition (n_grid), cell size (cell_size), standard deviation of presences (sd_p), standard deviation of absences (sd_a), Moran's I spatial autocorrelation (spa_auto), and environmental similarity based on euclidean distance (env_sim).

grid: A SpatRaster object with blocks

See also

Examples

if (FALSE) {
require(terra)
require(dplyr)

# Load datasets
data(spp)
f <- system.file("external/somevar.tif", package = "flexsdm")
somevar <- terra::rast(f)

# Lest practice with a single species
single_spp <- spp %>% dplyr::filter(species == "sp3")
part <- part_spat(
  env_layer = somevar,
  data = single_spp,
  x = "x",
  y = "y",
  pr_ab = "pr_ab",
  min_res_mult = 10,
  max_res_mult = 500,
  num_grids = 30,
  n_part = 2,
  prop = 0.5
)
part

part$part
part$best_part_info
part$grid

# Lets explore Grid object

plot(part$grid)
points(part$part[c("x", "y")],
       col = c("blue", "red")[part$part$.part],
       cex = 0.5,
       pch = 19
)

terra::res(part$grid)
terra::res(somevar)

# Note that is a layer with block partition, but it has a
# different resolution than the original environmental variables.
# In the case you wish have a layer with the same properties
# (i.e. resolution, extent, NAs) than your original environmental
# variables you can use the \code{\link{get_block}} function.

grid_env <- get_block(env_layer = somevar, best_grid = part$grid)

plot(grid_env) # this is a block layer with the same layer
# properties as environmental variables.
points(part$part[c("x", "y")],
       col = c("blue", "red")[part$part$.part],
       cex = 0.5,
       pch = 19
)
# This layer could be very useful in case you need sample
# pseudo_absence or background point
# See examples in \code{\link{backgroudp}} and \code{\link{pseudoabs}}


# Lest try with a higher number of partition
part <- part_spat(
  env_layer = somevar,
  data = single_spp,
  x = "x",
  y = "y",
  pr_ab = "pr_ab",
  min_res_mult = 10,
  max_res_mult = 500,
  num_grids = 30,
  n_part = 4,
  prop = 0.5
)

# Lets explore Grid object
plot(part$grid, col=gray.colors(4))
points(part$part[c("x", "y")],
       col = rainbow(n = 4)[part$part$.part],
       cex = 0.5,
       pch = 19
)


# Now lets learn use these functions with several species
spp2 <- split(spp, spp$species)
class(spp2)
length(spp2)
names(spp2)

part_list <- lapply(spp2, function(x) {
  result <- part_spat(
    env_layer = somevar,
    data = x,
    x = "x",
    y = "y",
    pr_ab = "pr_ab",
    min_res_mult = 10,
    max_res_mult = 500,
    num_grids = 30,
    n_part = 2,
    prop = 0.5
  )
  result
})

# Lets create a single database for all species
occ_part <- dplyr::bind_rows(lapply(
  part_list,
  function(x) x[[1]]
), .id = "species")
occ_part

# Lets get a the best grid info for all species
grid_info <- dplyr::bind_rows(lapply(
  part_list,
  function(x) x[[2]]
), .id = "species")

# Lets get a the best grid layer for all species
grid_layer2 <-
  lapply(grid_layer, function(x) {
    get_block(env_layer = somevar[[1]], best_grid = x)
  })
grid_layer2 <- terra::rast(grid_layer2)
grid_layer2
plot(grid_layer2)


# Block partition for presences-only database
single_spp <- spp %>%
  dplyr::filter(species == "sp2", pr_ab == 1)
single_spp
single_spp$pr_ab %>% unique() # only presences

part <- part_spat(
  env_layer = somevar,
  data = single_spp,
  x = "x",
  y = "y",
  pr_ab = "pr_ab",
  min_res_mult = 10,
  max_res_mult = 500,
  num_grids = 30,
  n_part = 2,
  prop = 0.5
)

part$part
part$best_part_info
part$grid

plot(part$grid)
points(
  part$part[c("x", "y")],
  col = c("blue", "red")[part$part$.part],
  cex = 0.5,
  pch = 19
)
}