Fit and validate Maximum Entropy models with exploration of hyper-parameters

Fit and validate Maximum Entropy models with exploration of hyper-parameters

tune_mx(
  data,
  response,
  predictors,
  predictors_f = NULL,
  background = NULL,
  partition,
  grid = NULL,
  thr = NULL,
  metric = "TSS",
  clamp = TRUE,
  pred_type = "cloglog",
  ...
)

Arguments

data
data.frame. Database with response (0,1) and predictors values.

response
character. Column name with species absence-presence data (0,1).

predictors
character. Vector with the column names of quantitative predictor variables (i.e. continuous or discrete variables). Usage predictors = c("aet", "cwd", "tmin")

predictors_f
character. Vector with the column names of qualitative predictor variables (i.e. ordinal or nominal variables type). Usage predictors_f = c("landform")

background
data.frame. Database with response column only with 0 and predictors variables. All column names must be consistent with data

partition
character. Column name with training and validation partition groups.

grid
data.frame. Provide a data frame object with algorithm hyper-parameters values to be tested. It Is recommended to generate this data.frame with grid() function. Hyper-parameters needed for tuning are 'regmult' and 'classes' (any combination of next letters l -linear-, q -quadratic-, h -hinge-, p -product-, and t -threshold-).

thr
character. Threshold used to get binary suitability values (i.e. 0,1) useful for threshold-dependent performance metrics. It is possible not define a threshold or use more than one, in these cases, function will return the best model and the best threshold. It is necessary to provide a vector for this argument. The next threshold area available:

lpt: The highest threshold at which there is no omission. Usage thr=c(type='lpt').

equal_sens_spec: Threshold at which the sensitivity and specificity are equal (aka threshold that maximizes the TSS).

max_sens_spec: Threshold at which the sum of the sensitivity and specificity is the highest. Usage thr=c(type='max_sens_spec').

max_kappa: The threshold at which Kappa is the highest ("max kappa"). Usage thr=c(type='max_kappa').

max_jaccard: The threshold at which Jaccard is the highest. Usage thr=c(type='max_jaccard').

max_sorensen: The threshold at which Sorensen is highest. Usage thr=c(type='max_sorensen').

max_fpb: The threshold at which FPB is highest. Usage thr=c(type='max_fpb').

specific: A threshold value specified by user. Usage thr=c(type='specific', sens='0.6'). 'sens' refers to models will be binarized using this suitability value.

metric
character. Performance metric used for selecting the best combination of hyper-parameter values. Can be used one of the next metrics SORENSEN, JACCARD, FPB, TSS, KAPPA, AUC, and BOYCE. TSS is used as default.

clamp
logical. It is set with TRUE, predictors and features are restricted to the range seen during model training.

pred_type
character. Type of response required available "link", "exponential", "cloglog" and "logistic". Default "cloglog"

...

Value

A list object with:

model: A "maxnet" "lognet" "glmnet" class object. This object can be used for predicting.

predictors: A character with quantitative (elements names with c) and qualitative (elements names with f) variables use for modeling.

tune_performance: Performance metric (see sdm_eval) for each combination of the hyper-parameters.

best_hyper_performance: Hyper-parameters values and performance metric (see sdm_eval) for the best hyper-parameters combination.

selected_thresholds: Value of the threshold selected.

all_thresholds: Value of all threshold.

See also