Fit and validate Neural Networks models with exploration of hyper-parameters

Fit and validate Neural Networks models with exploration of hyper-parameters

tune_net(
  data,
  response,
  predictors,
  predictors_f = NULL,
  fit_formula = NULL,
  partition,
  grid = NULL,
  thr = NULL,
  metric = "TSS"
)

Arguments

data
data.frame. Database with response (0,1) and predictors values.

response
character. Column name with species absence-presence data (0,1).

predictors
character. Vector with the column names of quantitative predictor variables (i.e. continuous variables). Usage predictors = c("aet", "cwd", "tmin")

predictors_f
character. Vector with the column names of qualitative predictor variables (i.e. ordinal or nominal variables type). Usage predictors_f = c("landform")

fit_formula
formula. A formula object with response and predictor variables (e.g. formula(pr_ab ~ aet + ppt_jja + pH + awc + depth + landform)). Note that the variables used here must be consistent with those used in response, predictors, and predictors_f arguments. Defaul NULL.

partition
character. Column name with training and validation partition groups.

grid
data.frame. Provide a data frame object with algorithm hyper-parameters values to be tested. It Is recommended to generate this data.frame with grid() function. Hyper-parameters needed for tuning are 'size' and 'decay'.

thr
character. Threshold used to get binary suitability values (i.e. 0,1). It is useful for threshold-dependent performance metrics. It is possible to use more than one threshold type. It is necessary to provide a vector for this argument. The next threshold area available:

lpt: The highest threshold at which there is no omission.

equal_sens_spec: Threshold at which the sensitivity and specificity are equal.

max_sens_spec: Threshold at which the sum of the sensitivity and specificity is the highest (aka threshold that maximizes the TSS).

max_jaccard: The threshold at which Jaccard is the highest.

max_sorensen: The threshold at which Sorensen is highest.

max_fpb: The threshold at which FPB is highest.

sensitivity: Threshold based on a specified sensitivity value. Usage thr = c('sensitivity', sens='0.6') or thr = c('sensitivity'). 'sens' refers to sensitivity value. If it is not specified a sensitivity values, function will use by default 0.9

In the case of use more than one threshold type it is necessary concatenate threshold types, e.g., thr=c('lpt', 'max_sens_spec', 'max_jaccard'), or thr=c('lpt', 'max_sens_spec', 'sensitivity', sens='0.8'), or thr=c('lpt', 'max_sens_spec', 'sensitivity'). Function will use all thresholds if no threshold is specified

metric
character. Performance metric used for selecting the best combination of hyper-parameter values. Can be used one of the next metrics SORENSEN, JACCARD, FPB, TSS, KAPPA, AUC, and BOYCE. TSS is used as default.

Value

A list object with:

model: A "nnet" class object. This object can be used for predicting.

predictors: A tibble with quantitative (c column names) and qualitative (f column names) variables use for modeling.

performance: Hyper-parameters values and performance metric (see sdm_eval) for the best hyper-parameters combination.

hyper_performance: Performance metric (see sdm_eval) for each combination of the hyper-parameters.

data_ens: Predicted suitability for each test partition based on the best model. This database is used in fit_ensemble

See also

Examples

# \dontrun{
data(abies)
abies

#> # A tibble: 1,400 x 13
#>       id pr_ab        x        y   aet   cwd  tmin ppt_djf ppt_jja    pH    awc
#>    <int> <dbl>    <dbl>    <dbl> <dbl> <dbl> <dbl>   <dbl>   <dbl> <dbl>  <dbl>
#>  1   715     0  -95417.  314240.  323.  546.  1.24    62.7   17.8   5.77 0.108 
#>  2  5680     0   98987. -159415.  448.  815.  9.43   130.     6.43  5.60 0.160 
#>  3  7907     0  121474.  -99463.  182.  271. -4.95   151.    11.2   0    0     
#>  4  1850     0  -39976.  -17456.  372.  946.  8.78   116.     2.70  6.41 0.0972
#>  5  1702     0  111372.  -91404.  209.  399. -4.03   165.     9.27  0    0     
#>  6 10036     0 -255715.  392229.  308.  535.  4.66   166.    16.5   5.70 0.0777
#>  7 12384     0 -311765.  380213.  568.  352.  4.38   480.    41.2   5.80 0.110 
#>  8  6513     0  111360. -120229.  327.  633.  4.93   163.     8.91  1.18 0.0116
#>  9  9884     0 -284326.  442136.  377.  446.  3.99   296.    16.8   5.96 0.0900
#> 10  8651     0  137640. -110538.  215.  265. -4.62   180.     9.57  0    0     
#> # ... with 1,390 more rows, and 2 more variables: depth <dbl>, landform <fct>


# We will partition the data with the k-fold method

abies2 <- part_random(
  data = abies,
  pr_ab = "pr_ab",
  method = c(method = "kfold", folds = 5)
)

# pr_ab columns is species presence and absences (i.e. the response variable)
# from aet to landform are the predictors variables (landform is a qualitative variable)

# Hyper-parameter values for tuning
tune_grid <-
  expand.grid(
    size = c(2, 4, 6, 8, 10),
    decay = c(0.001, 0.05, 0.1, 1, 3, 4, 5, 10)
  )

net_t <-
  tune_net(
    data = abies2,
    response = "pr_ab",
    predictors = c(
      "aet", "cwd", "tmin", "ppt_djf",
      "ppt_jja", "pH", "awc", "depth"
    ),
    predictors_f = c("landform"),
    partition = ".part",
    grid = tune_grid,
    thr = "max_sens_spec",
    metric = "TSS",
  )

#> Formula used for model fitting:
#> pr_ab ~ aet + cwd + tmin + ppt_djf + ppt_jja + pH + awc + depth + landform

#> Replica number: 1/1

#> Partition number: 1/5

#> Partition number: 2/5

#> Partition number: 3/5

#> Partition number: 4/5

#> Partition number: 5/5

#> Formula used for model fitting:
#> pr_ab ~ aet + cwd + tmin + ppt_djf + ppt_jja + pH + awc + depth + landform

#> Replica number: 1/1

#> Partition number: 1/5

#> Partition number: 2/5

#> Partition number: 3/5

#> Partition number: 4/5

#> Partition number: 5/5


# Outputs
net_t$model

#> a 22-8-1 network with 193 weights
#> inputs: aet cwd tmin ppt_djf ppt_jja pH awc depth landform2 landform3 landform4 landform5 landform6 landform7 landform8 landform9 landform10 landform11 landform12 landform13 landform14 landform15 
#> output(s): pr_ab 
#> options were - entropy fitting  decay=3

net_t$predictors

#> # A tibble: 1 x 9
#>   c1    c2    c3    c4      c5      c6    c7    c8    f       
#>   <chr> <chr> <chr> <chr>   <chr>   <chr> <chr> <chr> <chr>   
#> 1 aet   cwd   tmin  ppt_djf ppt_jja pH    awc   depth landform

net_t$performance

#> # A tibble: 1 x 27
#>    size decay model threshold   thr_value n_presences n_absences TPR_mean TPR_sd
#>   <dbl> <dbl> <chr> <chr>           <dbl>       <int>      <int>    <dbl>  <dbl>
#> 1     8     3 net   max_sens_s~     0.463         700        700    0.919 0.0486
#> # ... with 18 more variables: TNR_mean <dbl>, TNR_sd <dbl>,
#> #   SORENSEN_mean <dbl>, SORENSEN_sd <dbl>, JACCARD_mean <dbl>,
#> #   JACCARD_sd <dbl>, FPB_mean <dbl>, FPB_sd <dbl>, OR_mean <dbl>, OR_sd <dbl>,
#> #   TSS_mean <dbl>, TSS_sd <dbl>, AUC_mean <dbl>, AUC_sd <dbl>,
#> #   BOYCE_mean <dbl>, BOYCE_sd <dbl>, IMAE_mean <dbl>, IMAE_sd <dbl>

net_t$hyper_performance

#> # A tibble: 40 x 24
#>     size  decay model threshold    TPR_mean TPR_sd TNR_mean TNR_sd SORENSEN_mean
#>    <dbl>  <dbl> <chr> <chr>           <dbl>  <dbl>    <dbl>  <dbl>         <dbl>
#>  1     2  0.001 net   max_sens_sp~    0.4   0.548     0.6   0.548          0.267
#>  2     2  0.05  net   max_sens_sp~    0.926 0.0436    0.726 0.101          0.842
#>  3     2  0.1   net   max_sens_sp~    0.939 0.0314    0.687 0.0557         0.834
#>  4     2  1     net   max_sens_sp~    0.901 0.0333    0.8   0.0466         0.858
#>  5     2  3     net   max_sens_sp~    0.911 0.0297    0.764 0.0520         0.849
#>  6     2  4     net   max_sens_sp~    0.87  0.0796    0.796 0.101          0.839
#>  7     2  5     net   max_sens_sp~    0.864 0.0664    0.793 0.0657         0.834
#>  8     2 10     net   max_sens_sp~    0.904 0.0380    0.741 0.0402         0.836
#>  9     4  0.001 net   max_sens_sp~    0.933 0.0543    0.587 0.344          0.806
#> 10     4  0.05  net   max_sens_sp~    0.911 0.0306    0.76  0.0708         0.848
#> # ... with 30 more rows, and 15 more variables: SORENSEN_sd <dbl>,
#> #   JACCARD_mean <dbl>, JACCARD_sd <dbl>, FPB_mean <dbl>, FPB_sd <dbl>,
#> #   OR_mean <dbl>, OR_sd <dbl>, TSS_mean <dbl>, TSS_sd <dbl>, AUC_mean <dbl>,
#> #   AUC_sd <dbl>, BOYCE_mean <dbl>, BOYCE_sd <dbl>, IMAE_mean <dbl>,
#> #   IMAE_sd <dbl>

net_t$data_ens

#> # A tibble: 1,400 x 5
#>    rnames replicates part  pr_ab   pred
#>    <chr>  <chr>      <chr> <fct>  <dbl>
#>  1 4      .part      1     0     0.0129
#>  2 13     .part      1     0     0.0134
#>  3 19     .part      1     0     0.0884
#>  4 20     .part      1     0     0.415 
#>  5 24     .part      1     0     0.873 
#>  6 27     .part      1     0     0.223 
#>  7 28     .part      1     0     0.0128
#>  8 31     .part      1     0     0.0915
#>  9 42     .part      1     0     0.0877
#> 10 48     .part      1     0     0.0861
#> # ... with 1,390 more rows

# }