Fit and validate Generalized Additive Models based on Ensemble of Small of Model approach

This function constructs Generalized Additive Models using the Ensemble of Small Model (ESM) approach (Breiner et al., 2015, 2018).

esm_gam(data, response, predictors, partition, thr = NULL)

Arguments

data
data.frame. Database with the response (0,1) and predictors values.

response
character. Column name with species absence-presence data (0,1)

predictors
character. Vector with the column names of quantitative predictor variables (i.e. continuous variables). This function does not allow categorical variables Usage predictors = c("aet", "cwd", "tmin"). This function only can construct models with continuous variables.

partition
character. Column name with training and validation partition groups.

thr
character. Threshold used to get binary suitability values (i.e. 0,1). It is useful for threshold-dependent performance metrics. It is possible to use more than one threshold type. It is necessary to provide a vector for this argument. The next threshold area available:

equal_sens_spec: Threshold at which the sensitivity and specificity are equal.

max_sens_spec: Threshold at which the sum of the sensitivity and specificity is the highest (aka threshold that maximizes the TSS).

max_jaccard: The threshold at which Jaccard is the highest.

max_sorensen: The threshold at which Sorensen is highest.

max_fpb: The threshold at which FPB is highest.

sensitivity: Threshold based on a specified sensitivity value. Usage thr = c('sensitivity', sens='0.6') or thr = c('sensitivity'). 'sens' refers to sensitivity value. If it is not specified a sensitivity values, the function will use by default 0.9

If the user wants to include more than one threshold type, it is necessary to concatenate threshold types, e.g., thr=c('max_sens_spec', 'max_jaccard'), or thr=c('max_sens_spec', 'sensitivity', sens='0.8'), or thr=c('max_sens_spec', 'sensitivity'). Function will use all thresholds if no threshold is specified

Value

A list object with:

esm_model: A list with "Gam" class object for each bivariate model. This object can be used for predicting ensemble of small model with sdm_predict function.

predictors: A tibble with variables use for modeling.

performance: Performance metric (see sdm_eval). Those threshold dependent metric are calculated based on the threshold specified in thr argument.

Details

This method consists of creating bivariate models with all the pair-wise combinations of predictors and perform an ensemble based on the average of suitability weighted by Somers'D metric (D = 2 x (AUC -0.5)). ESM is recommended for modeling species with few occurrences. This function does not allow categorical variables because the use of these types of variables could be problematic when using with few occurrences. Further detail see Breiner et al. (2015, 2018)

References

Breiner, F. T., Guisan, A., Bergamini, A., & Nobis, M. P. (2015). Overcoming limitations of modelling rare species by using ensembles of small models. Methods in Ecology and Evolution, 6(10), 1210-218. https://doi.org/10.1111/2041-210X.12403

Breiner, F. T., Nobis, M. P., Bergamini, A., & Guisan, A. (2018). Optimizing ensembles of small models for predicting the distribution of species with few occurrences. Methods in Ecology and Evolution, 9(4), 802-808. https://doi.org/10.1111/2041-210X.12957

See also

Examples

# \dontrun{
data("abies")
require(dplyr)

# Using k-fold partition method
set.seed(10)
abies2 <- abies %>%
  na.omit() %>%
  group_by(pr_ab) %>%
  dplyr::slice_sample(n = 10) %>%
  group_by()

abies2 <- part_random(
  data = abies2,
  pr_ab = "pr_ab",
  method = c(method = "kfold", folds = 3)
)
abies2

#> # A tibble: 20 x 14
#>       id pr_ab        x        y   aet   cwd   tmin ppt_djf ppt_jja    pH    awc
#>    <int> <dbl>    <dbl>    <dbl> <dbl> <dbl>  <dbl>   <dbl>   <dbl> <dbl>  <dbl>
#>  1 12040     0 -308909.  384248.  573.  332.  4.84     521.   48.8   5.63 0.108 
#>  2 10361     0 -254286.  417158.  260.  469.  2.93     151.   15.1   6.20 0.0950
#>  3  9402     0 -286979.  386206.  587.  376.  6.45     333.   15.7   5.5  0.160 
#>  4  9815     0 -291849.  445595.  443.  455.  4.39     332.   19.1   6    0.0700
#>  5 10524     0 -256658.  184438.  355.  568.  5.87     303.   10.6   5.20 0.0800
#>  6  8860     0  121343. -164170.  354.  733.  3.97     182.    9.83  0    0     
#>  7  6431     0  107903. -122968.  461.  578.  4.87     161.    7.66  5.90 0.0900
#>  8 11730     0 -333903.  431238.  561.  364.  6.73     387.   25.2   5.80 0.130 
#>  9   808     0 -150163.  357180.  339.  564.  2.64     220.   15.3   6.40 0.100 
#> 10 11054     0 -293663.  340981.  477.  396.  3.89     332.   26.4   4.60 0.0634
#> 11  2960     1  -49273.  181752.  512.  275.  0.920    319.   17.3   5.92 0.0900
#> 12  3065     1  126907. -198892.  322.  544.  0.700    203.   10.6   5.60 0.110 
#> 13  5527     1  116751. -181089.  261.  537.  0.363    178.    7.43  0    0     
#> 14  4035     1  -31777.  115940.  394.  440.  2.07     298.   11.2   6.01 0.0769
#> 15  4081     1   -5158.   90159.  301.  502.  0.703    203.   14.6   6.11 0.0633
#> 16  3087     1  102151. -143976.  299.  425. -2.08     205.   13.4   3.88 0.110 
#> 17  3495     1  -19586.   89803.  438.  419.  2.13     189.   15.2   6.19 0.0959
#> 18  4441     1   49405.  -60502.  362.  582.  2.42     218.    7.84  5.64 0.0786
#> 19   301     1 -132516.  270845.  367.  196. -2.56     422.   26.3   6.70 0.0300
#> 20  3162     1   59905.  -53634.  319.  626.  1.99     212.    4.50  4.51 0.0396
#> # ... with 3 more variables: depth <dbl>, landform <fct>, .part <int>


# Without threshold specification and with kfold
esm_gam_t1 <- esm_gam(
  data = abies2,
  response = "pr_ab",
  predictors = c("aet", "cwd", "tmin", "ppt_djf", "ppt_jja", "pH", "awc", "depth"),
  partition = ".part",
  thr = NULL
)

#>   |                                                                              |                                                                      |   0%  |                                                                              |==                                                                    |   4%  |                                                                              |=====                                                                 |   7%  |                                                                              |========                                                              |  11%  |                                                                              |==========                                                            |  14%  |                                                                              |============                                                          |  18%  |                                                                              |===============                                                       |  21%  |                                                                              |==================                                                    |  25%  |                                                                              |====================                                                  |  29%  |                                                                              |======================                                                |  32%  |                                                                              |=========================                                             |  36%  |                                                                              |============================                                          |  39%  |                                                                              |==============================                                        |  43%  |                                                                              |================================                                      |  46%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================                                |  54%  |                                                                              |========================================                              |  57%  |                                                                              |==========================================                            |  61%  |                                                                              |=============================================                         |  64%  |                                                                              |================================================                      |  68%  |                                                                              |==================================================                    |  71%  |                                                                              |====================================================                  |  75%  |                                                                              |=======================================================               |  79%  |                                                                              |==========================================================            |  82%  |                                                                              |============================================================          |  86%  |                                                                              |==============================================================        |  89%  |                                                                              |=================================================================     |  93%  |                                                                              |====================================================================  |  96%  |                                                                              |======================================================================| 100%


esm_gam_t1$esm_model # bivariate model

#> $`0.259259259259259`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.99998 Residual
#> Residual Deviance: 9.86509 
#> 
#> $`0.875`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11 Residual
#> Residual Deviance: 0 
#> 
#> $`-0.398148148148148`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00021 Residual
#> Residual Deviance: 14.29457 
#> 
#> $`-0.1875`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.99981 Residual
#> Residual Deviance: 19.34354 
#> 
#> $`-0.814814814814815`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00026 Residual
#> Residual Deviance: 18.79893 
#> 
#> $`-0.402777777777778`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00046 Residual
#> Residual Deviance: 18.90059 
#> 
#> $`0.377314814814815`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.0002 Residual
#> Residual Deviance: 0.79746 
#> 
#> $`0.888888888888889`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00012 Residual
#> Residual Deviance: 0 
#> 
#> $`0.497685185185185`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.99978 Residual
#> Residual Deviance: 6.63526 
#> 
#> $`0.231481481481481`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00017 Residual
#> Residual Deviance: 9.81768 
#> 
#> $`-0.791666666666667`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.99999 Residual
#> Residual Deviance: 18.03072 
#> 
#> $`-0.75`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.99993 Residual
#> Residual Deviance: 18.02498 
#> 
#> $`0.273148148148148`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.99984 Residual
#> Residual Deviance: 1.61772 
#> 
#> $`0.875`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.9999 Residual
#> Residual Deviance: 0 
#> 
#> $`0.875`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.99993 Residual
#> Residual Deviance: 0 
#> 
#> $`0.916666666666667`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.9999 Residual
#> Residual Deviance: 1e-04 
#> 
#> $`1`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.99999 Residual
#> Residual Deviance: 0 
#> 
#> $`0.875`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00001 Residual
#> Residual Deviance: 0 
#> 
#> $`0.231481481481481`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00017 Residual
#> Residual Deviance: 16.04344 
#> 
#> $`-0.233796296296296`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00003 Residual
#> Residual Deviance: 17.36921 
#> 
#> $`-0.0787037037037037`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00001 Residual
#> Residual Deviance: 16.74941 
#> 
#> $`0.564814814814815`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00003 Residual
#> Residual Deviance: 6.57265 
#> 
#> $`-0.391203703703704`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00007 Residual
#> Residual Deviance: 20.28762 
#> 
#> $`-0.0416666666666666`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.99974 Residual
#> Residual Deviance: 18.81425 
#> 
#> $`0.569444444444444`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.99977 Residual
#> Residual Deviance: 2.74366 
#> 
#> $`-0.884259259259259`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00011 Residual
#> Residual Deviance: 19.40843 
#> 
#> $`0.0972222222222221`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.99984 Residual
#> Residual Deviance: 8.20297 
#> 
#> $`0.594907407407407`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.9997 Residual
#> Residual Deviance: 3.09105 
#> 

esm_gam_t1$predictors

#> # A tibble: 1 x 8
#>   c1    c2    c3    c4      c5      c6    c7    c8   
#>   <chr> <chr> <chr> <chr>   <chr>   <chr> <chr> <chr>
#> 1 aet   cwd   tmin  ppt_djf ppt_jja pH    awc   depth

esm_gam_t1$performance

#> # A tibble: 7 x 25
#>   model   threshold    thr_value n_presences n_absences TPR_mean TPR_sd TNR_mean
#>   <chr>   <chr>            <dbl>       <int>      <int>    <dbl>  <dbl>    <dbl>
#> 1 esm_gam equal_sens_~     0.767          10         10        1      0        1
#> 2 esm_gam lpt              0.767          10         10        1      0        1
#> 3 esm_gam max_fpb          0.767          10         10        1      0        1
#> 4 esm_gam max_jaccard      0.767          10         10        1      0        1
#> 5 esm_gam max_sens_sp~     0.767          10         10        1      0        1
#> 6 esm_gam max_sorensen     0.767          10         10        1      0        1
#> 7 esm_gam sensitivity      0.991          10         10        1      0        1
#> # ... with 17 more variables: TNR_sd <dbl>, SORENSEN_mean <dbl>,
#> #   SORENSEN_sd <dbl>, JACCARD_mean <dbl>, JACCARD_sd <dbl>, FPB_mean <dbl>,
#> #   FPB_sd <dbl>, OR_mean <dbl>, OR_sd <dbl>, TSS_mean <dbl>, TSS_sd <dbl>,
#> #   AUC_mean <dbl>, AUC_sd <dbl>, BOYCE_mean <dbl>, BOYCE_sd <dbl>,
#> #   IMAE_mean <dbl>, IMAE_sd <dbl>


# Test with rep_kfold partition
abies2 <- abies2 %>% select(-starts_with("."))

set.seed(10)
abies2 <- part_random(
  data = abies2,
  pr_ab = "pr_ab",
  method = c(method = "rep_kfold", folds = 3, replicates = 5)
)
abies2

#> # A tibble: 20 x 18
#>       id pr_ab        x        y   aet   cwd   tmin ppt_djf ppt_jja    pH    awc
#>    <int> <dbl>    <dbl>    <dbl> <dbl> <dbl>  <dbl>   <dbl>   <dbl> <dbl>  <dbl>
#>  1 12040     0 -308909.  384248.  573.  332.  4.84     521.   48.8   5.63 0.108 
#>  2 10361     0 -254286.  417158.  260.  469.  2.93     151.   15.1   6.20 0.0950
#>  3  9402     0 -286979.  386206.  587.  376.  6.45     333.   15.7   5.5  0.160 
#>  4  9815     0 -291849.  445595.  443.  455.  4.39     332.   19.1   6    0.0700
#>  5 10524     0 -256658.  184438.  355.  568.  5.87     303.   10.6   5.20 0.0800
#>  6  8860     0  121343. -164170.  354.  733.  3.97     182.    9.83  0    0     
#>  7  6431     0  107903. -122968.  461.  578.  4.87     161.    7.66  5.90 0.0900
#>  8 11730     0 -333903.  431238.  561.  364.  6.73     387.   25.2   5.80 0.130 
#>  9   808     0 -150163.  357180.  339.  564.  2.64     220.   15.3   6.40 0.100 
#> 10 11054     0 -293663.  340981.  477.  396.  3.89     332.   26.4   4.60 0.0634
#> 11  2960     1  -49273.  181752.  512.  275.  0.920    319.   17.3   5.92 0.0900
#> 12  3065     1  126907. -198892.  322.  544.  0.700    203.   10.6   5.60 0.110 
#> 13  5527     1  116751. -181089.  261.  537.  0.363    178.    7.43  0    0     
#> 14  4035     1  -31777.  115940.  394.  440.  2.07     298.   11.2   6.01 0.0769
#> 15  4081     1   -5158.   90159.  301.  502.  0.703    203.   14.6   6.11 0.0633
#> 16  3087     1  102151. -143976.  299.  425. -2.08     205.   13.4   3.88 0.110 
#> 17  3495     1  -19586.   89803.  438.  419.  2.13     189.   15.2   6.19 0.0959
#> 18  4441     1   49405.  -60502.  362.  582.  2.42     218.    7.84  5.64 0.0786
#> 19   301     1 -132516.  270845.  367.  196. -2.56     422.   26.3   6.70 0.0300
#> 20  3162     1   59905.  -53634.  319.  626.  1.99     212.    4.50  4.51 0.0396
#> # ... with 7 more variables: depth <dbl>, landform <fct>, .part1 <int>,
#> #   .part2 <int>, .part3 <int>, .part4 <int>, .part5 <int>


esm_gam_t2 <- esm_gam(
  data = abies2,
  response = "pr_ab",
  predictors = c("aet", "cwd", "tmin", "ppt_djf", "ppt_jja", "pH", "awc", "depth"),
  partition = ".part",
  thr = NULL
)

#>   |                                                                              |                                                                      |   0%  |                                                                              |==                                                                    |   4%  |                                                                              |=====                                                                 |   7%  |                                                                              |========                                                              |  11%  |                                                                              |==========                                                            |  14%  |                                                                              |============                                                          |  18%  |                                                                              |===============                                                       |  21%  |                                                                              |==================                                                    |  25%  |                                                                              |====================                                                  |  29%  |                                                                              |======================                                                |  32%  |                                                                              |=========================                                             |  36%  |                                                                              |============================                                          |  39%  |                                                                              |==============================                                        |  43%  |                                                                              |================================                                      |  46%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================                                |  54%  |                                                                              |========================================                              |  57%  |                                                                              |==========================================                            |  61%  |                                                                              |=============================================                         |  64%  |                                                                              |================================================                      |  68%  |                                                                              |==================================================                    |  71%  |                                                                              |====================================================                  |  75%  |                                                                              |=======================================================               |  79%  |                                                                              |==========================================================            |  82%  |                                                                              |============================================================          |  86%  |                                                                              |==============================================================        |  89%  |                                                                              |=================================================================     |  93%  |                                                                              |====================================================================  |  96%  |                                                                              |======================================================================| 100%

esm_gam_t2$esm_model # bivariate model

#> $`0.150925925925926`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.99998 Residual
#> Residual Deviance: 9.86509 
#> 
#> $`0.917592592592593`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11 Residual
#> Residual Deviance: 0 
#> 
#> $`0.2125`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00021 Residual
#> Residual Deviance: 14.29457 
#> 
#> $`-0.0745370370370371`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.99981 Residual
#> Residual Deviance: 19.34354 
#> 
#> $`-0.236111111111111`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00026 Residual
#> Residual Deviance: 18.79893 
#> 
#> $`-0.358333333333333`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00046 Residual
#> Residual Deviance: 18.90059 
#> 
#> $`0.455555555555555`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.0002 Residual
#> Residual Deviance: 0.79746 
#> 
#> $`0.838425925925926`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00012 Residual
#> Residual Deviance: 0 
#> 
#> $`0.308796296296296`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.99978 Residual
#> Residual Deviance: 6.63526 
#> 
#> $`0.318981481481482`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00017 Residual
#> Residual Deviance: 9.81768 
#> 
#> $`-0.267592592592593`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.99999 Residual
#> Residual Deviance: 18.03072 
#> 
#> $`-0.271296296296296`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.99993 Residual
#> Residual Deviance: 18.02498 
#> 
#> $`0.619907407407407`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.99984 Residual
#> Residual Deviance: 1.61772 
#> 
#> $`0.816666666666667`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.9999 Residual
#> Residual Deviance: 0 
#> 
#> $`0.930555555555556`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.99993 Residual
#> Residual Deviance: 0 
#> 
#> $`0.851851851851852`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.9999 Residual
#> Residual Deviance: 1e-04 
#> 
#> $`0.925`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.99999 Residual
#> Residual Deviance: 0 
#> 
#> $`0.851851851851852`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00001 Residual
#> Residual Deviance: 0 
#> 
#> $`0.271759259259259`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00017 Residual
#> Residual Deviance: 16.04344 
#> 
#> $`0`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00003 Residual
#> Residual Deviance: 17.36921 
#> 
#> $`-0.0310185185185186`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00001 Residual
#> Residual Deviance: 16.74941 
#> 
#> $`0.277777777777778`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00003 Residual
#> Residual Deviance: 6.57265 
#> 
#> $`-0.0314814814814814`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00007 Residual
#> Residual Deviance: 20.28762 
#> 
#> $`-0.0337962962962963`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.99974 Residual
#> Residual Deviance: 18.81425 
#> 
#> $`0.339351851851852`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.99977 Residual
#> Residual Deviance: 2.74366 
#> 
#> $`-0.380555555555556`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00011 Residual
#> Residual Deviance: 19.40843 
#> 
#> $`0.311111111111111`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.99984 Residual
#> Residual Deviance: 8.20297 
#> 
#> $`0.568518518518518`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.9997 Residual
#> Residual Deviance: 3.09105 
#> 

esm_gam_t2$predictors

#> # A tibble: 1 x 8
#>   c1    c2    c3    c4      c5      c6    c7    c8   
#>   <chr> <chr> <chr> <chr>   <chr>   <chr> <chr> <chr>
#> 1 aet   cwd   tmin  ppt_djf ppt_jja pH    awc   depth

esm_gam_t2$performance

#> # A tibble: 7 x 25
#>   model   threshold    thr_value n_presences n_absences TPR_mean TPR_sd TNR_mean
#>   <chr>   <chr>            <dbl>       <int>      <int>    <dbl>  <dbl>    <dbl>
#> 1 esm_gam equal_sens_~     0.505          50         50    0.961  0.104    1    
#> 2 esm_gam lpt              0.415          50         50    1      0        0.961
#> 3 esm_gam max_fpb          0.415          50         50    1      0        0.961
#> 4 esm_gam max_jaccard      0.415          50         50    1      0        0.961
#> 5 esm_gam max_sens_sp~     0.415          50         50    0.961  0.104    1    
#> 6 esm_gam max_sorensen     0.415          50         50    1      0        0.961
#> 7 esm_gam sensitivity      0.711          50         50    1      0        0.961
#> # ... with 17 more variables: TNR_sd <dbl>, SORENSEN_mean <dbl>,
#> #   SORENSEN_sd <dbl>, JACCARD_mean <dbl>, JACCARD_sd <dbl>, FPB_mean <dbl>,
#> #   FPB_sd <dbl>, OR_mean <dbl>, OR_sd <dbl>, TSS_mean <dbl>, TSS_sd <dbl>,
#> #   AUC_mean <dbl>, AUC_sd <dbl>, BOYCE_mean <dbl>, BOYCE_sd <dbl>,
#> #   IMAE_mean <dbl>, IMAE_sd <dbl>


# Test with other bootstrap
abies2 <- abies2 %>% select(-starts_with("."))

set.seed(10)
abies2 <- part_random(
  data = abies2,
  pr_ab = "pr_ab",
  method = c(method = "boot", replicates = 10, proportion = 0.7)
)
abies2

#> # A tibble: 20 x 23
#>       id pr_ab        x        y   aet   cwd   tmin ppt_djf ppt_jja    pH    awc
#>    <int> <dbl>    <dbl>    <dbl> <dbl> <dbl>  <dbl>   <dbl>   <dbl> <dbl>  <dbl>
#>  1 12040     0 -308909.  384248.  573.  332.  4.84     521.   48.8   5.63 0.108 
#>  2 10361     0 -254286.  417158.  260.  469.  2.93     151.   15.1   6.20 0.0950
#>  3  9402     0 -286979.  386206.  587.  376.  6.45     333.   15.7   5.5  0.160 
#>  4  9815     0 -291849.  445595.  443.  455.  4.39     332.   19.1   6    0.0700
#>  5 10524     0 -256658.  184438.  355.  568.  5.87     303.   10.6   5.20 0.0800
#>  6  8860     0  121343. -164170.  354.  733.  3.97     182.    9.83  0    0     
#>  7  6431     0  107903. -122968.  461.  578.  4.87     161.    7.66  5.90 0.0900
#>  8 11730     0 -333903.  431238.  561.  364.  6.73     387.   25.2   5.80 0.130 
#>  9   808     0 -150163.  357180.  339.  564.  2.64     220.   15.3   6.40 0.100 
#> 10 11054     0 -293663.  340981.  477.  396.  3.89     332.   26.4   4.60 0.0634
#> 11  2960     1  -49273.  181752.  512.  275.  0.920    319.   17.3   5.92 0.0900
#> 12  3065     1  126907. -198892.  322.  544.  0.700    203.   10.6   5.60 0.110 
#> 13  5527     1  116751. -181089.  261.  537.  0.363    178.    7.43  0    0     
#> 14  4035     1  -31777.  115940.  394.  440.  2.07     298.   11.2   6.01 0.0769
#> 15  4081     1   -5158.   90159.  301.  502.  0.703    203.   14.6   6.11 0.0633
#> 16  3087     1  102151. -143976.  299.  425. -2.08     205.   13.4   3.88 0.110 
#> 17  3495     1  -19586.   89803.  438.  419.  2.13     189.   15.2   6.19 0.0959
#> 18  4441     1   49405.  -60502.  362.  582.  2.42     218.    7.84  5.64 0.0786
#> 19   301     1 -132516.  270845.  367.  196. -2.56     422.   26.3   6.70 0.0300
#> 20  3162     1   59905.  -53634.  319.  626.  1.99     212.    4.50  4.51 0.0396
#> # ... with 12 more variables: depth <dbl>, landform <fct>, .part1 <chr>,
#> #   .part2 <chr>, .part3 <chr>, .part4 <chr>, .part5 <chr>, .part6 <chr>,
#> #   .part7 <chr>, .part8 <chr>, .part9 <chr>, .part10 <chr>


esm_gam_t3 <- esm_gam(
  data = abies2,
  response = "pr_ab",
  predictors = c("aet", "cwd", "tmin", "ppt_djf", "ppt_jja", "pH", "awc", "depth"),
  partition = ".part",
  thr = NULL
)

#>   |                                                                              |                                                                      |   0%  |                                                                              |==                                                                    |   4%  |                                                                              |=====                                                                 |   7%  |                                                                              |========                                                              |  11%  |                                                                              |==========                                                            |  14%  |                                                                              |============                                                          |  18%  |                                                                              |===============                                                       |  21%  |                                                                              |==================                                                    |  25%  |                                                                              |====================                                                  |  29%  |                                                                              |======================                                                |  32%  |                                                                              |=========================                                             |  36%  |                                                                              |============================                                          |  39%  |                                                                              |==============================                                        |  43%  |                                                                              |================================                                      |  46%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================                                |  54%  |                                                                              |========================================                              |  57%  |                                                                              |==========================================                            |  61%  |                                                                              |=============================================                         |  64%  |                                                                              |================================================                      |  68%  |                                                                              |==================================================                    |  71%  |                                                                              |====================================================                  |  75%  |                                                                              |=======================================================               |  79%  |                                                                              |==========================================================            |  82%  |                                                                              |============================================================          |  86%  |                                                                              |==============================================================        |  89%  |                                                                              |=================================================================     |  93%  |                                                                              |====================================================================  |  96%  |                                                                              |======================================================================| 100%

esm_gam_t3$esm_model # bivariate model

#> $`0.577777777777778`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.99998 Residual
#> Residual Deviance: 9.86509 
#> 
#> $`0.933333333333333`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11 Residual
#> Residual Deviance: 0 
#> 
#> $`0.566666666666667`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00021 Residual
#> Residual Deviance: 14.29457 
#> 
#> $`0.633333333333333`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.99981 Residual
#> Residual Deviance: 19.34354 
#> 
#> $`0.733333333333333`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00026 Residual
#> Residual Deviance: 18.79893 
#> 
#> $`0.711111111111111`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00046 Residual
#> Residual Deviance: 18.90059 
#> 
#> $`0.944444444444444`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.0002 Residual
#> Residual Deviance: 0.79746 
#> 
#> $`0.933333333333333`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00012 Residual
#> Residual Deviance: 0 
#> 
#> $`0.666666666666667`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.99978 Residual
#> Residual Deviance: 6.63526 
#> 
#> $`0.733333333333333`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00017 Residual
#> Residual Deviance: 9.81768 
#> 
#> $`0.333333333333333`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.99999 Residual
#> Residual Deviance: 18.03072 
#> 
#> $`0.422222222222222`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.99993 Residual
#> Residual Deviance: 18.02498 
#> 
#> $`0.955555555555555`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.99984 Residual
#> Residual Deviance: 1.61772 
#> 
#> $`0.922222222222222`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.9999 Residual
#> Residual Deviance: 0 
#> 
#> $`0.955555555555555`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.99993 Residual
#> Residual Deviance: 0 
#> 
#> $`0.966666666666667`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.9999 Residual
#> Residual Deviance: 1e-04 
#> 
#> $`1`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.99999 Residual
#> Residual Deviance: 0 
#> 
#> $`0.955555555555555`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00001 Residual
#> Residual Deviance: 0 
#> 
#> $`0.811111111111111`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00017 Residual
#> Residual Deviance: 16.04344 
#> 
#> $`0.644444444444444`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00003 Residual
#> Residual Deviance: 17.36921 
#> 
#> $`0.6`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00001 Residual
#> Residual Deviance: 16.74941 
#> 
#> $`0.866666666666667`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00003 Residual
#> Residual Deviance: 6.57265 
#> 
#> $`0.622222222222222`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00007 Residual
#> Residual Deviance: 20.28762 
#> 
#> $`0.644444444444444`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.99974 Residual
#> Residual Deviance: 18.81425 
#> 
#> $`0.811111111111111`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.99977 Residual
#> Residual Deviance: 2.74366 
#> 
#> $`0.477777777777778`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 11.00011 Residual
#> Residual Deviance: 19.40843 
#> 
#> $`0.811111111111111`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.99984 Residual
#> Residual Deviance: 8.20297 
#> 
#> $`0.955555555555555`
#> Call:
#> gam::gam(formula = formula1, family = "binomial", data = data)
#> 
#> Degrees of Freedom: 19 total; 10.9997 Residual
#> Residual Deviance: 3.09105 
#> 

esm_gam_t3$predictors

#> # A tibble: 1 x 8
#>   c1    c2    c3    c4      c5      c6    c7    c8   
#>   <chr> <chr> <chr> <chr>   <chr>   <chr> <chr> <chr>
#> 1 aet   cwd   tmin  ppt_djf ppt_jja pH    awc   depth

esm_gam_t3$performance

#> # A tibble: 7 x 25
#>   model   threshold    thr_value n_presences n_absences TPR_mean TPR_sd TNR_mean
#>   <chr>   <chr>            <dbl>       <int>      <int>    <dbl>  <dbl>    <dbl>
#> 1 esm_gam equal_sens_~     0.484          30         30    0.867  0.172    1    
#> 2 esm_gam lpt              0.429          30         30    1      0        0.867
#> 3 esm_gam max_fpb          0.429          30         30    1      0        0.867
#> 4 esm_gam max_jaccard      0.429          30         30    1      0        0.867
#> 5 esm_gam max_sens_sp~     0.484          30         30    0.867  0.172    1    
#> 6 esm_gam max_sorensen     0.429          30         30    1      0        0.867
#> 7 esm_gam sensitivity      0.565          30         30    1      0        0.867
#> # ... with 17 more variables: TNR_sd <dbl>, SORENSEN_mean <dbl>,
#> #   SORENSEN_sd <dbl>, JACCARD_mean <dbl>, JACCARD_sd <dbl>, FPB_mean <dbl>,
#> #   FPB_sd <dbl>, OR_mean <dbl>, OR_sd <dbl>, TSS_mean <dbl>, TSS_sd <dbl>,
#> #   AUC_mean <dbl>, AUC_sd <dbl>, BOYCE_mean <dbl>, BOYCE_sd <dbl>,
#> #   IMAE_mean <dbl>, IMAE_sd <dbl>

# }