Fit and validate Maximum Entropy models

Fit and validate Maximum Entropy models

fit_max(
  data,
  response,
  predictors,
  predictors_f = NULL,
  fit_formula = NULL,
  partition,
  background = NULL,
  thr = NULL,
  clamp = TRUE,
  classes = "default",
  pred_type = "cloglog",
  regmult = 1
)

Arguments

data
data.frame. Database with response (0,1) and predictors values.

response
character. Column name with species absence-presence data (0,1).

predictors
character. Vector with the column names of quantitative predictor variables (i.e. continuous variables). Usage predictors = c("aet", "cwd", "tmin")

predictors_f
character. Vector with the column names of qualitative predictor variables (i.e. ordinal or nominal variables type). Usage predictors_f = c("landform")

fit_formula
formula. A formula object with response and predictor variables see maxnet.formula function from maxnet package. Note that the variables used here must be consistent with those used in response, predictors, and predictors_f arguments. Default NULL.

partition
character. Column name with training and validation partition groups.

background
data.frame. Database with response column only with 0 and predictors variables. All column names must be consistent with data. Default NULL

thr
character. Threshold used to get binary suitability values (i.e. 0,1). It is useful for threshold-dependent performance metrics. It is possible to use more than one threshold type. It is necessary to provide a vector for this argument. The next threshold area available:

lpt: The highest threshold at which there is no omission.

equal_sens_spec: Threshold at which the sensitivity and specificity are equal.

max_sens_spec: Threshold at which the sum of the sensitivity and specificity is the highest (aka threshold that maximizes the TSS).

max_jaccard: The threshold at which Jaccard is the highest.

max_sorensen: The threshold at which Sorensen is highest.

max_fpb: The threshold at which FPB is highest.

sensitivity: Threshold based on a specified sensitivity value. Usage thr = c('sensitivity', sens='0.6') or thr = c('sensitivity'). 'sens' refers to sensitivity value. If it is not specified a sensitivity values, function will use by default 0.9

In the case of use more than one threshold type it is necessary concatenate threshold types, e.g., thr=c('lpt', 'max_sens_spec', 'max_jaccard'), or thr=c('lpt', 'max_sens_spec', 'sensitivity', sens='0.8'), or thr=c('lpt', 'max_sens_spec', 'sensitivity'). Function will use all thresholds if no threshold is specified

clamp
logical. It is set with TRUE, predictors and features are restricted to the range seen during model training.

classes
character. A single feature of any combinations of them. Features are symbolized by letters: l (linear), q (quadratic), h (hinge), p (product), and t (threshold). Usage classes = "lpq". Default "default" (see details).

pred_type
character. Type of response required available "link", "exponential", "cloglog" and "logistic". Default "cloglog"

regmult
numeric. A constant to adjust regularization. Default 1.

Value

A list object with:

model: A "MaxEnt" class object. This object can be used for predicting.

predictors: A tibble with quantitative (c column names) and qualitative (f column names) variables use for modeling.

performance: Performance metric (see sdm_eval). Those threshold dependent metric are calculated based on the threshold specified in thr argument.

data_ens: Predicted suitability for each test partition based on the best model. This database is used in fit_ensemble

Details

EXPLAIN HERE DEFAUL SELECTION OF FEATURES BASED ON number of occurrences if (np < 10) classes <- "l" else if (np < 15) classes <- "lq" else if (np < 80) classes <- "lqh"

Examples

# \dontrun{
data("abies")
data("backg")
abies # environmental conditions of presence-absence data

#> # A tibble: 1,400 x 13
#>       id pr_ab        x        y   aet   cwd  tmin ppt_djf ppt_jja    pH    awc
#>    <int> <dbl>    <dbl>    <dbl> <dbl> <dbl> <dbl>   <dbl>   <dbl> <dbl>  <dbl>
#>  1   715     0  -95417.  314240.  323.  546.  1.24    62.7   17.8   5.77 0.108 
#>  2  5680     0   98987. -159415.  448.  815.  9.43   130.     6.43  5.60 0.160 
#>  3  7907     0  121474.  -99463.  182.  271. -4.95   151.    11.2   0    0     
#>  4  1850     0  -39976.  -17456.  372.  946.  8.78   116.     2.70  6.41 0.0972
#>  5  1702     0  111372.  -91404.  209.  399. -4.03   165.     9.27  0    0     
#>  6 10036     0 -255715.  392229.  308.  535.  4.66   166.    16.5   5.70 0.0777
#>  7 12384     0 -311765.  380213.  568.  352.  4.38   480.    41.2   5.80 0.110 
#>  8  6513     0  111360. -120229.  327.  633.  4.93   163.     8.91  1.18 0.0116
#>  9  9884     0 -284326.  442136.  377.  446.  3.99   296.    16.8   5.96 0.0900
#> 10  8651     0  137640. -110538.  215.  265. -4.62   180.     9.57  0    0     
#> # ... with 1,390 more rows, and 2 more variables: depth <dbl>, landform <fct>

backg # environmental conditions of background points

#> # A tibble: 5,000 x 13
#>    pr_ab        x        y   aet   cwd  tmin ppt_djf ppt_jja    pH     awc depth
#>    <dbl>    <dbl>    <dbl> <dbl> <dbl> <dbl>   <dbl>   <dbl> <dbl>   <dbl> <dbl>
#>  1     0  160779. -449968.  280. 1137. 13.5     71.3    1.19 0     0         0  
#>  2     0   36849.   24152.  260.  382. -3.17   171.    17.5  0.212 0.00347 201  
#>  3     0 -240171.   90032.  400.  700.  8.68   285.     5.02 5.72  0.0804   50.1
#>  4     0 -152421. -143518.  367.  843.  9.01    72.0    1.20 7.54  0.170   154. 
#>  5     0 -193191.   24152.  397.  842.  8.97   125.     1.98 6.20  0.131   122. 
#>  6     0 -277971.  223682.  385.  637.  4.93   226.     8.16 5.81  0.0512   56.2
#>  7     0 -313341.  270122.  582.  406.  6.29   334.    18.4  5.80  0.168   201  
#>  8     0   54399.  -15538.  346.  195. -5.22   142.    13.0  5.60  0.120   201  
#>  9     0  282549. -582268.  285. 1097. 11.4     56.3    1.39 6.57  0.135    68.4
#> 10     0  104079. -178618.  385.  871.  6.76   147.     7.80 6.10  0.0300   41  
#> # ... with 4,990 more rows, and 2 more variables: percent_clay <dbl>,
#> #   landform <fct>


# Using k-fold partition method
# Remember that the partition method, number of folds or replications must
# be the same for presence-absence and background points datasets
abies2 <- part_random(
  data = abies,
  pr_ab = "pr_ab",
  method = c(method = "kfold", folds = 5)
)
abies2

#> # A tibble: 1,400 x 14
#>       id pr_ab        x        y   aet   cwd  tmin ppt_djf ppt_jja    pH    awc
#>    <int> <dbl>    <dbl>    <dbl> <dbl> <dbl> <dbl>   <dbl>   <dbl> <dbl>  <dbl>
#>  1   715     0  -95417.  314240.  323.  546.  1.24    62.7   17.8   5.77 0.108 
#>  2  5680     0   98987. -159415.  448.  815.  9.43   130.     6.43  5.60 0.160 
#>  3  7907     0  121474.  -99463.  182.  271. -4.95   151.    11.2   0    0     
#>  4  1850     0  -39976.  -17456.  372.  946.  8.78   116.     2.70  6.41 0.0972
#>  5  1702     0  111372.  -91404.  209.  399. -4.03   165.     9.27  0    0     
#>  6 10036     0 -255715.  392229.  308.  535.  4.66   166.    16.5   5.70 0.0777
#>  7 12384     0 -311765.  380213.  568.  352.  4.38   480.    41.2   5.80 0.110 
#>  8  6513     0  111360. -120229.  327.  633.  4.93   163.     8.91  1.18 0.0116
#>  9  9884     0 -284326.  442136.  377.  446.  3.99   296.    16.8   5.96 0.0900
#> 10  8651     0  137640. -110538.  215.  265. -4.62   180.     9.57  0    0     
#> # ... with 1,390 more rows, and 3 more variables: depth <dbl>, landform <fct>,
#> #   .part <int>


backg2 <- part_random(
  data = backg,
  pr_ab = "pr_ab",
  method = c(method = "kfold", folds = 5)
)
backg2

#> # A tibble: 5,000 x 14
#>    pr_ab        x        y   aet   cwd  tmin ppt_djf ppt_jja    pH     awc depth
#>    <dbl>    <dbl>    <dbl> <dbl> <dbl> <dbl>   <dbl>   <dbl> <dbl>   <dbl> <dbl>
#>  1     0  160779. -449968.  280. 1137. 13.5     71.3    1.19 0     0         0  
#>  2     0   36849.   24152.  260.  382. -3.17   171.    17.5  0.212 0.00347 201  
#>  3     0 -240171.   90032.  400.  700.  8.68   285.     5.02 5.72  0.0804   50.1
#>  4     0 -152421. -143518.  367.  843.  9.01    72.0    1.20 7.54  0.170   154. 
#>  5     0 -193191.   24152.  397.  842.  8.97   125.     1.98 6.20  0.131   122. 
#>  6     0 -277971.  223682.  385.  637.  4.93   226.     8.16 5.81  0.0512   56.2
#>  7     0 -313341.  270122.  582.  406.  6.29   334.    18.4  5.80  0.168   201  
#>  8     0   54399.  -15538.  346.  195. -5.22   142.    13.0  5.60  0.120   201  
#>  9     0  282549. -582268.  285. 1097. 11.4     56.3    1.39 6.57  0.135    68.4
#> 10     0  104079. -178618.  385.  871.  6.76   147.     7.80 6.10  0.0300   41  
#> # ... with 4,990 more rows, and 3 more variables: percent_clay <dbl>,
#> #   landform <fct>, .part <int>


max_t1 <- fit_max(
  data = abies2,
  response = "pr_ab",
  predictors = c("aet", "ppt_jja", "pH", "awc", "depth"),
  predictors_f = c("landform"),
  partition = ".part",
  background = backg2,
  thr = c("max_sens_spec", "equal_sens_spec", "max_sorensen"),
  clamp = TRUE,
  classes = "default",
  pred_type = "cloglog",
  regmult = 1
)

#> Formula used for model fitting:
#> ~aet + ppt_jja + pH + awc + depth + I(aet^2) + I(ppt_jja^2) + I(pH^2) + I(awc^2) + I(depth^2) + hinge(aet) + hinge(ppt_jja) + hinge(pH) + hinge(awc) + hinge(depth) + ppt_jja:aet + pH:aet + awc:aet + depth:aet + pH:ppt_jja + awc:ppt_jja + depth:ppt_jja + awc:pH + depth:pH + depth:awc - 1

#> Replica number: 1/1

#> Partition number: 1/5

#> Partition number: 2/5

#> Partition number: 3/5

#> Partition number: 4/5

#> Partition number: 5/5

length(max_t1)

#> [1] 4

max_t1$model

#> 
#> Call:  glmnet::glmnet(x = mm, y = as.factor(p), family = "binomial",      weights = weights, lambda = 10^(seq(4, 0, length.out = 200)) *          sum(reg)/length(reg) * sum(p)/sum(weights), standardize = F,      penalty.factor = reg) 
#> 
#>     Df  %Dev  Lambda
#> 1    0  0.00 159.200
#> 2    0  0.00 152.000
#> 3    0  0.00 145.200
#> 4    0  0.00 138.600
#> 5    0  0.00 132.300
#> 6    0  0.00 126.300
#> 7    0  0.00 120.600
#> 8    0  0.00 115.200
#> 9    0  0.00 110.000
#> 10   0  0.00 105.000
#> 11   0  0.00 100.200
#> 12   0  0.00  95.710
#> 13   0  0.00  91.380
#> 14   0  0.00  87.250
#> 15   0  0.00  83.300
#> 16   0  0.00  79.530
#> 17   0  0.00  75.940
#> 18   0  0.00  72.500
#> 19   0  0.00  69.220
#> 20   0  0.00  66.090
#> 21   0  0.00  63.100
#> 22   0  0.00  60.250
#> 23   0  0.00  57.520
#> 24   0  0.00  54.920
#> 25   0  0.00  52.440
#> 26   0  0.00  50.070
#> 27   0  0.00  47.800
#> 28   0  0.00  45.640
#> 29   0  0.00  43.580
#> 30   0  0.00  41.610
#> 31   0  0.00  39.720
#> 32   0  0.00  37.930
#> 33   0  0.00  36.210
#> 34   0  0.00  34.570
#> 35   0  0.00  33.010
#> 36   0  0.00  31.520
#> 37   0  0.00  30.090
#> 38   0  0.00  28.730
#> 39   0  0.00  27.430
#> 40   0  0.00  26.190
#> 41   0  0.00  25.010
#> 42   0  0.00  23.870
#> 43   0  0.00  22.800
#> 44   0  0.00  21.760
#> 45   0  0.00  20.780
#> 46   0  0.00  19.840
#> 47   0  0.00  18.940
#> 48   0  0.00  18.090
#> 49   0  0.00  17.270
#> 50   0  0.00  16.490
#> 51   0  0.00  15.740
#> 52   0  0.00  15.030
#> 53   0  0.00  14.350
#> 54   0  0.00  13.700
#> 55   0  0.00  13.080
#> 56   0  0.00  12.490
#> 57   0  0.00  11.920
#> 58   0  0.00  11.390
#> 59   0  0.00  10.870
#> 60   0  0.00  10.380
#> 61   0  0.00   9.909
#> 62   0  0.00   9.461
#> 63   0  0.00   9.033
#> 64   0  0.00   8.624
#> 65   0  0.00   8.234
#> 66   0  0.00   7.862
#> 67   0  0.00   7.506
#> 68   0  0.00   7.167
#> 69   0  0.00   6.843
#> 70   0  0.00   6.533
#> 71   0  0.00   6.238
#> 72   0  0.00   5.956
#> 73   0  0.00   5.686
#> 74   0  0.00   5.429
#> 75   0  0.00   5.184
#> 76   0  0.00   4.949
#> 77   0  0.00   4.725
#> 78   0  0.00   4.512
#> 79   0  0.00   4.307
#> 80   0  0.00   4.113
#> 81   0  0.00   3.927
#> 82   0  0.00   3.749
#> 83   0  0.00   3.579
#> 84   0  0.00   3.418
#> 85   0  0.00   3.263
#> 86   0  0.00   3.115
#> 87   0  0.00   2.975
#> 88   0  0.00   2.840
#> 89   0  0.00   2.712
#> 90   0  0.00   2.589
#> 91   0  0.00   2.472
#> 92   0  0.00   2.360
#> 93   0  0.00   2.253
#> 94   1  0.18   2.151
#> 95   1  0.41   2.054
#> 96   1  0.61   1.961
#> 97   1  0.80   1.872
#> 98   1  0.97   1.788
#> 99   1  1.13   1.707
#> 100  1  1.28   1.630
#> 101  1  1.41   1.556
#> 102  1  1.54   1.486
#> 103  1  1.65   1.418
#> 104  1  1.75   1.354
#> 105  1  1.85   1.293
#> 106  3  2.03   1.235
#> 107  3  2.36   1.179
#> 108  3  2.66   1.125
#> 109  3  2.94   1.075
#> 110  3  3.20   1.026
#> 111  3  3.44   0.980
#> 112  3  3.66   0.935
#> 113  3  3.87   0.893
#> 114  3  4.06   0.853
#> 115  3  4.23   0.814
#> 116  3  4.39   0.777
#> 117  3  4.55   0.742
#> 118  3  4.69   0.708
#> 119  4  4.88   0.676
#> 120  4  5.10   0.646
#> 121  4  5.31   0.617
#> 122  4  5.50   0.589
#> 123  4  5.67   0.562
#> 124  4  5.83   0.537
#> 125  4  5.98   0.512
#> 126  5  6.17   0.489
#> 127  5  6.41   0.467
#> 128  5  6.64   0.446
#> 129  5  6.86   0.426
#> 130  5  7.06   0.406
#> 131  6  7.27   0.388
#> 132  7  7.47   0.371
#> 133  7  7.66   0.354
#> 134  7  7.84   0.338
#> 135  7  8.01   0.322
#> 136  7  8.17   0.308
#> 137  7  8.31   0.294
#> 138  8  8.45   0.281
#> 139  8  8.59   0.268
#> 140  9  8.80   0.256
#> 141  9  9.03   0.244
#> 142  9  9.26   0.233
#> 143 11  9.47   0.223
#> 144  9  9.66   0.213
#> 145  9  9.84   0.203
#> 146  9 10.01   0.194
#> 147  9 10.17   0.185
#> 148  9 10.31   0.177
#> 149 10 10.47   0.169
#> 150 10 10.62   0.161
#> 151 12 10.76   0.154
#> 152 11 10.90   0.147
#> 153 11 11.01   0.140
#> 154 10 11.12   0.134
#> 155 10 11.21   0.128
#> 156 10 11.30   0.122
#> 157 10 11.38   0.116
#> 158 12 11.46   0.111
#> 159 12 11.54   0.106
#> 160 12 11.61   0.101
#> 161 11 11.68   0.097
#> 162 10 11.74   0.092
#> 163 10 11.79   0.088
#> 164 11 11.84   0.084
#> 165 11 11.88   0.080
#> 166 12 11.95   0.077
#> 167 13 12.01   0.073
#> 168 13 12.08   0.070
#> 169 14 12.14   0.067
#> 170 15 12.21   0.064
#> 171 16 12.29   0.061
#> 172 16 12.36   0.058
#> 173 17 12.42   0.056
#> 174 17 12.49   0.053
#> 175 17 12.55   0.051
#> 176 18 12.61   0.048
#> 177 17 12.65   0.046
#> 178 17 12.69   0.044
#> 179 19 12.73   0.042
#> 180 19 12.80   0.040
#> 181 19 12.87   0.038
#> 182 20 12.96   0.037
#> 183 20 13.04   0.035
#> 184 21 13.12   0.033
#> 185 22 13.21   0.032
#> 186 23 13.63   0.030
#> 187 21 13.84   0.029
#> 188 21 13.98   0.028
#> 189 21 14.10   0.026
#> 190 25 14.24   0.025
#> 191 26 14.50   0.024
#> 192 28 14.79   0.023
#> 193 28 15.11   0.022
#> 194 28 15.40   0.021
#> 195 27 15.65   0.020
#> 196 27 15.88   0.019
#> 197 29 16.10   0.018
#> 198 29 16.32   0.017
#> 199 30 16.51   0.017
#> 200 31 16.70   0.016

max_t1$predictors

#> # A tibble: 1 x 6
#>   c1    c2      c3    c4    c5    f       
#>   <chr> <chr>   <chr> <chr> <chr> <chr>   
#> 1 aet   ppt_jja pH    awc   depth landform

max_t1$performance

#> # A tibble: 3 x 25
#>   model threshold      thr_value n_presences n_absences TPR_mean TPR_sd TNR_mean
#>   <chr> <chr>              <dbl>       <int>      <int>    <dbl>  <dbl>    <dbl>
#> 1 max   equal_sens_sp~     0.472         700        700    0.664 0.0354    0.666
#> 2 max   max_sens_spec      0.563         700        700    0.871 0.0440    0.549
#> 3 max   max_sorensen       0.273         700        700    0.921 0.0346    0.487
#> # ... with 17 more variables: TNR_sd <dbl>, SORENSEN_mean <dbl>,
#> #   SORENSEN_sd <dbl>, JACCARD_mean <dbl>, JACCARD_sd <dbl>, FPB_mean <dbl>,
#> #   FPB_sd <dbl>, OR_mean <dbl>, OR_sd <dbl>, TSS_mean <dbl>, TSS_sd <dbl>,
#> #   AUC_mean <dbl>, AUC_sd <dbl>, BOYCE_mean <dbl>, BOYCE_sd <dbl>,
#> #   IMAE_mean <dbl>, IMAE_sd <dbl>

max_t1$data_ens

#> # A tibble: 1,400 x 5
#>    rnames replicates part  pr_ab    pred
#>    <chr>  <chr>      <chr> <dbl>   <dbl>
#>  1 16     .part      1         0 0.168  
#>  2 17     .part      1         0 0.512  
#>  3 23     .part      1         0 0.602  
#>  4 30     .part      1         0 0.00898
#>  5 31     .part      1         0 0.321  
#>  6 34     .part      1         0 0.194  
#>  7 38     .part      1         0 0.0804 
#>  8 42     .part      1         0 0.145  
#>  9 43     .part      1         0 0.984  
#> 10 48     .part      1         0 0.174  
#> # ... with 1,390 more rows

# }