Perform geographical filtering on species occurrences

Perform geographical filtering on species occurrences

occfilt_geo(data, x, y, id, env_layer, method, prj = NULL)

Arguments

data
data.frame. Data.frame or tibble object with presences (or presence-absence) records, and coordinates

x
character. Column name with longitude data

y
character. Column name with latitude data

id
character. Column names with rows id. It is important that each row has its own unique code.

env_layer
SpatRaster. Raster variables that will be used to fit the model

method
character. Method to perform geographical thinning. Pair of points are filtered based on a geographical distance criteria.The following methods are available:

moran: Points are filetered based on the smallest distance that reduces Moran's I to values lower than 0.1. Latlong = T if occurrences are in a gegraphical projection. Usage method: method = c('moran').

cellsize: Points are filetered based on the environmental variables resolution. The factor can be specified. Usage method: method = c('cellsize', factor = '2').

defined: Points are filetered based on a provided distance value in km. Usage method: method = c('defined', d = 300).

prj
character. Projection string (PROJ4) for occurrences. Not necessary if the projection is WGS84.

Value

A tibble object with filtered data

Examples

if (FALSE) {
require(terra)
require(dplyr)

# Environmental variables
somevar <- system.file("external/somevar.tif", package = "flexsdm")
somevar <- terra::rast(somevar)

plot(somevar)

# Species occurrences
data("spp")
spp
spp1 <- spp %>% dplyr::filter(species == "sp1", pr_ab == 1)

somevar[[1]] %>% plot()
points(spp1 %>% select(x, y))

# Moran
filtered_1 <- occfilt_geo(
  data = spp1,
  x = "x",
  y = "y",
  id = "species",
  env_layer = somevar,
  method = c("moran"),
  prj = crs(somevar, proj = T)
)

# Cellsize
filtered_2 <- occfilt_geo(
  data = spp1,
  x = "x",
  y = "y",
  id = "species",
  env_layer = somevar,
  method = c("cellsize", factor = "3"),
  prj = crs(somevar, proj = T)
)

# Defined
filtered_3 <- occfilt_geo(
  data = spp1,
  x = "x",
  y = "y",
  id = "species",
  env_layer = somevar,
  method = c("defined", d = "30"),
  prj = crs(somevar, proj = T)
)
}