Extract environmental data based on x and y coordinates

Extract environmental data based on x and y coordinates

sdm_extract(data, x, y, env_layer, variables = NULL, filter_na = TRUE)

Arguments

data
data.frame. Database with presences, presence-absence, or pseudo-absence, records with x and y coordinates

x
character. Column name with longitude data

y
character. Column name with latitude data

env_layer
SpatRaster. Raster with environmental variables.

variables
character. Vector with the variable names of predictor variables Usage variables. = c("aet", "cwd", "tmin"). If no variable is specified, function will return data for all layers. Default NULL

filter_na
logical. If filter_na = TRUE (default), the rows with NA values for any of the environmental variables are removed from the returned tibble.

Value

A tibble that returns the original data base with additional columns for the extracted environmental variables at each xy location from the SpatRaster object used in 'env_layer'

Examples

# \dontrun{
require(terra)

# Load datasets
data(spp)
f <- system.file("external/somevar.tif", package = "flexsdm")
somevar <- terra::rast(f)

# Extract environmental data from somevar for all species locations in spp
ex_spp <-
  sdm_extract(
    data = spp,
    x = "x",
    y = "y",
    env_layer = somevar,
    variables = NULL,
    filter_na = FALSE
  )

# Extract environmental for two variables and remove rows with NAs
ex_spp2 <-
  sdm_extract(
    data = spp,
    x = "x",
    y = "y",
    env_layer = somevar,
    variables = c("CFP_3", "CFP_4"),
    filter_na = TRUE
  )

#> 35 rows were excluded from database because NAs were found


ex_spp

#> # A tibble: 1,150 x 8
#>    species        x        y pr_ab CFP_1 CFP_2 CFP_3 CFP_4
#>    <chr>      <dbl>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 sp1       -5541. -145138.     0  245. 1160.  24.9  9.72
#>  2 sp1      -51981.   16322.     0  434.  865.  23.4  8.27
#>  3 sp1     -269871.   69512.     1  470.  746.  21.6  7.80
#>  4 sp1      -96261.  -32008.     0  332.  980.  23.8  9.67
#>  5 sp1      269589. -566338.     0  276. 1091.  23.1 12.0 
#>  6 sp1       29829. -328468.     0  206. 1189.  23.5  7.84
#>  7 sp1     -152691.  393782.     0  381.  409.  13.2  2.06
#>  8 sp1     -195081.  253652.     0  354.  830.  23.8  9.64
#>  9 sp1        -951. -277978.     0  240. 1136.  23.1 10.2 
#> 10 sp1      145929. -271498.     0  237. 1031.  19.9  7.43
#> # ... with 1,140 more rows

ex_spp2

#> # A tibble: 1,115 x 6
#>    species        x        y pr_ab CFP_3 CFP_4
#>    <chr>      <dbl>    <dbl> <dbl> <dbl> <dbl>
#>  1 sp1       -5541. -145138.     0  24.9  9.72
#>  2 sp1      -51981.   16322.     0  23.4  8.27
#>  3 sp1     -269871.   69512.     1  21.6  7.80
#>  4 sp1      -96261.  -32008.     0  23.8  9.67
#>  5 sp1      269589. -566338.     0  23.1 12.0 
#>  6 sp1       29829. -328468.     0  23.5  7.84
#>  7 sp1     -152691.  393782.     0  13.2  2.06
#>  8 sp1     -195081.  253652.     0  23.8  9.64
#>  9 sp1        -951. -277978.     0  23.1 10.2 
#> 10 sp1      145929. -271498.     0  19.9  7.43
#> # ... with 1,105 more rows

# }