Merge model performance tables

Merge model performance tables

sdm_summarize(models)

Arguments

models
list of one or more models fitted with fit_ or tune_ functions, or a fit_ensemble output, a esm_ family function output. A list a single or several models fitted with some of fit_ or tune_ functions or object returned by the fit_ensemble function.

Value

Combined model performance table for all input models. Models fit with tune will include model performance for the best hyperparameters.

Examples

# \dontrun{
data(abies)
abies

#> # A tibble: 1,400 x 13
#>       id pr_ab        x        y   aet   cwd  tmin ppt_djf ppt_jja    pH    awc
#>    <int> <dbl>    <dbl>    <dbl> <dbl> <dbl> <dbl>   <dbl>   <dbl> <dbl>  <dbl>
#>  1   715     0  -95417.  314240.  323.  546.  1.24    62.7   17.8   5.77 0.108 
#>  2  5680     0   98987. -159415.  448.  815.  9.43   130.     6.43  5.60 0.160 
#>  3  7907     0  121474.  -99463.  182.  271. -4.95   151.    11.2   0    0     
#>  4  1850     0  -39976.  -17456.  372.  946.  8.78   116.     2.70  6.41 0.0972
#>  5  1702     0  111372.  -91404.  209.  399. -4.03   165.     9.27  0    0     
#>  6 10036     0 -255715.  392229.  308.  535.  4.66   166.    16.5   5.70 0.0777
#>  7 12384     0 -311765.  380213.  568.  352.  4.38   480.    41.2   5.80 0.110 
#>  8  6513     0  111360. -120229.  327.  633.  4.93   163.     8.91  1.18 0.0116
#>  9  9884     0 -284326.  442136.  377.  446.  3.99   296.    16.8   5.96 0.0900
#> 10  8651     0  137640. -110538.  215.  265. -4.62   180.     9.57  0    0     
#> # ... with 1,390 more rows, and 2 more variables: depth <dbl>, landform <fct>


# We will partition the data with the k-fold method

abies2 <- part_random(
  data = abies,
  pr_ab = "pr_ab",
  method = c(method = "kfold", folds = 5)
)

# Build a generalized additive model using fit_gam

gam_t1 <- fit_gam(
  data = abies2,
  response = "pr_ab",
  predictors = c("aet", "ppt_jja", "pH", "awc", "depth"),
  predictors_f = c("landform"),
  partition = ".part",
  thr = c("max_sens_spec", "equal_sens_spec", "max_sorensen")
)

#> Formula used for model fitting:
#> pr_ab ~ s(aet) + s(ppt_jja) + s(pH) + s(awc) + s(depth) + landform


gam_t1$performance

#> # A tibble: 3 x 25
#>   model threshold      thr_value n_presences n_absences TPR_mean TPR_sd TNR_mean
#>   <chr> <chr>              <dbl>       <int>      <int>    <dbl>  <dbl>    <dbl>
#> 1 gam   equal_sens_sp~     0.539         700        700    0.719 0.0430    0.719
#> 2 gam   max_sens_spec      0.436         700        700    0.801 0.0980    0.677
#> 3 gam   max_sorensen       0.379         700        700    0.903 0.0667    0.56 
#> # ... with 17 more variables: TNR_sd <dbl>, SORENSEN_mean <dbl>,
#> #   SORENSEN_sd <dbl>, JACCARD_mean <dbl>, JACCARD_sd <dbl>, FPB_mean <dbl>,
#> #   FPB_sd <dbl>, OR_mean <dbl>, OR_sd <dbl>, TSS_mean <dbl>, TSS_sd <dbl>,
#> #   AUC_mean <dbl>, AUC_sd <dbl>, BOYCE_mean <dbl>, BOYCE_sd <dbl>,
#> #   IMAE_mean <dbl>, IMAE_sd <dbl>


# Build a generalized linear model using fit_glm

glm_t1 <- fit_glm(
  data = abies2,
  response = "pr_ab",
  predictors = c("aet", "ppt_jja", "pH", "awc", "depth"),
  predictors_f = c("landform"),
  partition = ".part",
  thr = c("max_sens_spec", "equal_sens_spec", "max_sorensen"),
  poly = 0,
  inter_order = 0
)

#> Formula used for model fitting:
#> pr_ab ~ aet + ppt_jja + pH + awc + depth + landform

#> Replica number: 1/1

#> Partition number: 1/5

#> Partition number: 2/5

#> Partition number: 3/5

#> Partition number: 4/5

#> Partition number: 5/5


glm_t1$performance

#> # A tibble: 3 x 25
#>   model threshold      thr_value n_presences n_absences TPR_mean TPR_sd TNR_mean
#>   <chr> <chr>              <dbl>       <int>      <int>    <dbl>  <dbl>    <dbl>
#> 1 glm   equal_sens_sp~     0.523         700        700    0.663 0.0583    0.663
#> 2 glm   max_sens_spec      0.463         700        700    0.803 0.111     0.583
#> 3 glm   max_sorensen       0.356         700        700    0.876 0.0436    0.49 
#> # ... with 17 more variables: TNR_sd <dbl>, SORENSEN_mean <dbl>,
#> #   SORENSEN_sd <dbl>, JACCARD_mean <dbl>, JACCARD_sd <dbl>, FPB_mean <dbl>,
#> #   FPB_sd <dbl>, OR_mean <dbl>, OR_sd <dbl>, TSS_mean <dbl>, TSS_sd <dbl>,
#> #   AUC_mean <dbl>, AUC_sd <dbl>, BOYCE_mean <dbl>, BOYCE_sd <dbl>,
#> #   IMAE_mean <dbl>, IMAE_sd <dbl>


# Build a tuned random forest model using tune_raf

tune_grid <-
  expand.grid(mtry = seq(1, 7, 1))

rf_t1 <-
  tune_raf(
    data = abies2,
    response = "pr_ab",
    predictors = c(
      "aet", "cwd", "tmin", "ppt_djf",
      "ppt_jja", "pH", "awc", "depth"
    ),
    predictors_f = c("landform"),
    partition = ".part",
    grid = tune_grid,
    thr = c("max_sens_spec", "equal_sens_spec", "max_sorensen"),
    metric = "TSS",
  )

#> Formula used for model fitting:
#> pr_ab ~ aet + cwd + tmin + ppt_djf + ppt_jja + pH + awc + depth + landform

#> Replica number: 1/1

#> Partition number: 1/5

#> Partition number: 2/5

#> Partition number: 3/5

#> Partition number: 4/5

#> Partition number: 5/5

#> Formula used for model fitting:
#> pr_ab ~ aet + cwd + tmin + ppt_djf + ppt_jja + pH + awc + depth + landform

#> Replica number: 1/1

#> Partition number: 1/5

#> Partition number: 2/5

#> Partition number: 3/5

#> Partition number: 4/5

#> Partition number: 5/5


rf_t1$performance

#> # A tibble: 1 x 26
#>    mtry model threshold     thr_value n_presences n_absences TPR_mean TPR_sd
#>   <dbl> <chr> <chr>             <dbl>       <int>      <int>    <dbl>  <dbl>
#> 1     3 raf   max_sens_spec     0.606         700        700    0.916 0.0430
#> # ... with 18 more variables: TNR_mean <dbl>, TNR_sd <dbl>,
#> #   SORENSEN_mean <dbl>, SORENSEN_sd <dbl>, JACCARD_mean <dbl>,
#> #   JACCARD_sd <dbl>, FPB_mean <dbl>, FPB_sd <dbl>, OR_mean <dbl>, OR_sd <dbl>,
#> #   TSS_mean <dbl>, TSS_sd <dbl>, AUC_mean <dbl>, AUC_sd <dbl>,
#> #   BOYCE_mean <dbl>, BOYCE_sd <dbl>, IMAE_mean <dbl>, IMAE_sd <dbl>


# Merge sdm performance tables

merge_df <- sdm_summarize(models = list(gam_t1, glm_t1, rf_t1))

merge_df

#> # A tibble: 7 x 27
#>   model_ID model threshold      thr_value n_presences n_absences TPR_mean TPR_sd
#>      <int> <chr> <chr>              <dbl>       <int>      <int>    <dbl>  <dbl>
#> 1        1 gam   equal_sens_sp~     0.539         700        700    0.719 0.0430
#> 2        1 gam   max_sens_spec      0.436         700        700    0.801 0.0980
#> 3        1 gam   max_sorensen       0.379         700        700    0.903 0.0667
#> 4        2 glm   equal_sens_sp~     0.523         700        700    0.663 0.0583
#> 5        2 glm   max_sens_spec      0.463         700        700    0.803 0.111 
#> 6        2 glm   max_sorensen       0.356         700        700    0.876 0.0436
#> 7        3 raf   max_sens_spec      0.606         700        700    0.916 0.0430
#> # ... with 19 more variables: TNR_mean <dbl>, TNR_sd <dbl>,
#> #   SORENSEN_mean <dbl>, SORENSEN_sd <dbl>, JACCARD_mean <dbl>,
#> #   JACCARD_sd <dbl>, FPB_mean <dbl>, FPB_sd <dbl>, OR_mean <dbl>, OR_sd <dbl>,
#> #   TSS_mean <dbl>, TSS_sd <dbl>, AUC_mean <dbl>, AUC_sd <dbl>,
#> #   BOYCE_mean <dbl>, BOYCE_sd <dbl>, IMAE_mean <dbl>, IMAE_sd <dbl>,
#> #   mtry <dbl>

# }