Fit and validate Maximum Entropy models with exploration of hyper-parameters

Fit and validate Maximum Entropy models with exploration of hyper-parameters

tune_max(
  data,
  response,
  predictors,
  predictors_f = NULL,
  background = NULL,
  partition,
  grid = NULL,
  thr = NULL,
  metric = "TSS",
  clamp = TRUE,
  pred_type = "cloglog"
)

Arguments

data
data.frame. Database with response (0,1) and predictors values.

response
character. Column name with species absence-presence data (0,1).

predictors
character. Vector with the column names of quantitative predictor variables (i.e. continuous variables). Usage predictors = c("aet", "cwd", "tmin")

predictors_f
character. Vector with the column names of qualitative predictor variables (i.e. ordinal or nominal variables type). Usage predictors_f = c("landform")

background
data.frame. Database with response column only with 0 and predictors variables. All column names must be consistent with data

partition
character. Column name with training and validation partition groups.

grid
data.frame. Provide a data frame object with algorithm hyper-parameters values to be tested. It Is recommended to generate this data.frame with grid() function. Hyper-parameters needed for tuning are 'regmult' and 'classes' (any combination of next letters l -linear-, q -quadratic-, h -hinge-, p -product-, and t -threshold-).

thr
character. Threshold used to get binary suitability values (i.e. 0,1). It is useful for threshold-dependent performance metrics. It is possible to use more than one threshold type. It is necessary to provide a vector for this argument. The next threshold area available:

lpt: The highest threshold at which there is no omission.

equal_sens_spec: Threshold at which the sensitivity and specificity are equal.

max_sens_spec: Threshold at which the sum of the sensitivity and specificity is the highest (aka threshold that maximizes the TSS).

max_jaccard: The threshold at which Jaccard is the highest.

max_sorensen: The threshold at which Sorensen is highest.

max_fpb: The threshold at which FPB is highest.

sensitivity: Threshold based on a specified sensitivity value. Usage thr = c('sensitivity', sens='0.6') or thr = c('sensitivity'). 'sens' refers to sensitivity value. If it is not specified a sensitivity values, function will use by default 0.9

In the case of use more than one threshold type it is necessary concatenate threshold types, e.g., thr=c('lpt', 'max_sens_spec', 'max_jaccard'), or thr=c('lpt', 'max_sens_spec', 'sensitivity', sens='0.8'), or thr=c('lpt', 'max_sens_spec', 'sensitivity'). Function will use all thresholds if no threshold is specified

metric
character. Performance metric used for selecting the best combination of hyper-parameter values. Can be used one of the next metrics SORENSEN, JACCARD, FPB, TSS, KAPPA, AUC, and BOYCE. TSS is used as default.

clamp
logical. It is set with TRUE, predictors and features are restricted to the range seen during model training.

pred_type
character. Type of response required available "link", "exponential", "cloglog" and "logistic". Default "cloglog"

Value

A list object with:

model: A "maxnet" "lognet" "glmnet" class object. This object can be used for predicting.

predictors: A tibble with quantitative (c column names) and qualitative (f column names) variables use for modeling.

performance: Hyper-parameters values and performance metric (see sdm_eval) for the best hyper-parameters combination.

hyper_performance: Performance metric (see sdm_eval) for each combination of the hyper-parameters.

data_ens: Predicted suitability for each test partition based on the best model. This database is used in fit_ensemble

See also

Examples

# \dontrun{
data("abies")
data("backg")
abies # environmental conditions of presence-absence data

#> # A tibble: 1,400 x 13
#>       id pr_ab        x        y   aet   cwd  tmin ppt_djf ppt_jja    pH    awc
#>    <int> <dbl>    <dbl>    <dbl> <dbl> <dbl> <dbl>   <dbl>   <dbl> <dbl>  <dbl>
#>  1   715     0  -95417.  314240.  323.  546.  1.24    62.7   17.8   5.77 0.108 
#>  2  5680     0   98987. -159415.  448.  815.  9.43   130.     6.43  5.60 0.160 
#>  3  7907     0  121474.  -99463.  182.  271. -4.95   151.    11.2   0    0     
#>  4  1850     0  -39976.  -17456.  372.  946.  8.78   116.     2.70  6.41 0.0972
#>  5  1702     0  111372.  -91404.  209.  399. -4.03   165.     9.27  0    0     
#>  6 10036     0 -255715.  392229.  308.  535.  4.66   166.    16.5   5.70 0.0777
#>  7 12384     0 -311765.  380213.  568.  352.  4.38   480.    41.2   5.80 0.110 
#>  8  6513     0  111360. -120229.  327.  633.  4.93   163.     8.91  1.18 0.0116
#>  9  9884     0 -284326.  442136.  377.  446.  3.99   296.    16.8   5.96 0.0900
#> 10  8651     0  137640. -110538.  215.  265. -4.62   180.     9.57  0    0     
#> # ... with 1,390 more rows, and 2 more variables: depth <dbl>, landform <fct>

backg # environmental conditions of background points

#> # A tibble: 5,000 x 13
#>    pr_ab        x        y   aet   cwd  tmin ppt_djf ppt_jja    pH     awc depth
#>    <dbl>    <dbl>    <dbl> <dbl> <dbl> <dbl>   <dbl>   <dbl> <dbl>   <dbl> <dbl>
#>  1     0  160779. -449968.  280. 1137. 13.5     71.3    1.19 0     0         0  
#>  2     0   36849.   24152.  260.  382. -3.17   171.    17.5  0.212 0.00347 201  
#>  3     0 -240171.   90032.  400.  700.  8.68   285.     5.02 5.72  0.0804   50.1
#>  4     0 -152421. -143518.  367.  843.  9.01    72.0    1.20 7.54  0.170   154. 
#>  5     0 -193191.   24152.  397.  842.  8.97   125.     1.98 6.20  0.131   122. 
#>  6     0 -277971.  223682.  385.  637.  4.93   226.     8.16 5.81  0.0512   56.2
#>  7     0 -313341.  270122.  582.  406.  6.29   334.    18.4  5.80  0.168   201  
#>  8     0   54399.  -15538.  346.  195. -5.22   142.    13.0  5.60  0.120   201  
#>  9     0  282549. -582268.  285. 1097. 11.4     56.3    1.39 6.57  0.135    68.4
#> 10     0  104079. -178618.  385.  871.  6.76   147.     7.80 6.10  0.0300   41  
#> # ... with 4,990 more rows, and 2 more variables: percent_clay <dbl>,
#> #   landform <fct>


# Using k-fold partition method
# Remember that the partition method, number of folds or replications must
# be the same for presence-absence and background points datasets
abies2 <- part_random(
  data = abies,
  pr_ab = "pr_ab",
  method = c(method = "kfold", folds = 3)
)
abies2

#> # A tibble: 1,400 x 14
#>       id pr_ab        x        y   aet   cwd  tmin ppt_djf ppt_jja    pH    awc
#>    <int> <dbl>    <dbl>    <dbl> <dbl> <dbl> <dbl>   <dbl>   <dbl> <dbl>  <dbl>
#>  1   715     0  -95417.  314240.  323.  546.  1.24    62.7   17.8   5.77 0.108 
#>  2  5680     0   98987. -159415.  448.  815.  9.43   130.     6.43  5.60 0.160 
#>  3  7907     0  121474.  -99463.  182.  271. -4.95   151.    11.2   0    0     
#>  4  1850     0  -39976.  -17456.  372.  946.  8.78   116.     2.70  6.41 0.0972
#>  5  1702     0  111372.  -91404.  209.  399. -4.03   165.     9.27  0    0     
#>  6 10036     0 -255715.  392229.  308.  535.  4.66   166.    16.5   5.70 0.0777
#>  7 12384     0 -311765.  380213.  568.  352.  4.38   480.    41.2   5.80 0.110 
#>  8  6513     0  111360. -120229.  327.  633.  4.93   163.     8.91  1.18 0.0116
#>  9  9884     0 -284326.  442136.  377.  446.  3.99   296.    16.8   5.96 0.0900
#> 10  8651     0  137640. -110538.  215.  265. -4.62   180.     9.57  0    0     
#> # ... with 1,390 more rows, and 3 more variables: depth <dbl>, landform <fct>,
#> #   .part <int>


backg2 <- part_random(
  data = backg,
  pr_ab = "pr_ab",
  method = c(method = "kfold", folds = 3)
)
backg

#> # A tibble: 5,000 x 13
#>    pr_ab        x        y   aet   cwd  tmin ppt_djf ppt_jja    pH     awc depth
#>    <dbl>    <dbl>    <dbl> <dbl> <dbl> <dbl>   <dbl>   <dbl> <dbl>   <dbl> <dbl>
#>  1     0  160779. -449968.  280. 1137. 13.5     71.3    1.19 0     0         0  
#>  2     0   36849.   24152.  260.  382. -3.17   171.    17.5  0.212 0.00347 201  
#>  3     0 -240171.   90032.  400.  700.  8.68   285.     5.02 5.72  0.0804   50.1
#>  4     0 -152421. -143518.  367.  843.  9.01    72.0    1.20 7.54  0.170   154. 
#>  5     0 -193191.   24152.  397.  842.  8.97   125.     1.98 6.20  0.131   122. 
#>  6     0 -277971.  223682.  385.  637.  4.93   226.     8.16 5.81  0.0512   56.2
#>  7     0 -313341.  270122.  582.  406.  6.29   334.    18.4  5.80  0.168   201  
#>  8     0   54399.  -15538.  346.  195. -5.22   142.    13.0  5.60  0.120   201  
#>  9     0  282549. -582268.  285. 1097. 11.4     56.3    1.39 6.57  0.135    68.4
#> 10     0  104079. -178618.  385.  871.  6.76   147.     7.80 6.10  0.0300   41  
#> # ... with 4,990 more rows, and 2 more variables: percent_clay <dbl>,
#> #   landform <fct>



gridtest <-
  expand.grid(
    regmult = seq(0.1, 3, 0.5),
    classes = c("l", "lq", "lqhpt")
  )

max_t1 <- tune_max(
  data = abies2,
  response = "pr_ab",
  predictors = c("aet", "ppt_jja", "pH", "awc", "depth"),
  predictors_f = c("landform"),
  partition = ".part",
  background = backg2,
  grid = gridtest,
  thr = "max_sens_spec",
  metric = "TSS",
  clamp = TRUE,
  pred_type = "cloglog"
)

#> Replica number: 1/1

#> Partition number: 1/3

#> Partition number: 2/3

#> Partition number: 3/3

#> Formula used for model fitting:
#> ~aet + ppt_jja + pH + awc + depth + I(aet^2) + I(ppt_jja^2) + I(pH^2) + I(awc^2) + I(depth^2) + categorical(landform) - 1

#> Replica number: 1/1

#> Partition number: 1/3

#> Partition number: 2/3

#> Partition number: 3/3


length(max_t1)

#> [1] 5

max_t1$model

#> 
#> Call:  glmnet::glmnet(x = mm, y = as.factor(p), family = "binomial",      weights = weights, lambda = 10^(seq(4, 0, length.out = 200)) *          sum(reg)/length(reg) * sum(p)/sum(weights), standardize = F,      penalty.factor = reg) 
#> 
#>     Df  %Dev  Lambda
#> 1    0  0.00 1462.00
#> 2    0  0.00 1395.00
#> 3    0  0.00 1332.00
#> 4    0  0.00 1272.00
#> 5    0  0.00 1215.00
#> 6    0  0.00 1160.00
#> 7    0  0.00 1107.00
#> 8    0  0.00 1057.00
#> 9    0  0.00 1009.00
#> 10   0  0.00  963.60
#> 11   0  0.00  920.10
#> 12   0  0.00  878.40
#> 13   0  0.00  838.70
#> 14   0  0.00  800.80
#> 15   0  0.00  764.60
#> 16   0  0.00  730.00
#> 17   0  0.00  697.00
#> 18   0  0.00  665.40
#> 19   0  0.00  635.40
#> 20   0  0.00  606.60
#> 21   0  0.00  579.20
#> 22   0  0.00  553.00
#> 23   0  0.00  528.00
#> 24   0  0.00  504.10
#> 25   0  0.00  481.30
#> 26   0  0.00  459.50
#> 27   0  0.00  438.70
#> 28   0  0.00  418.90
#> 29   0  0.00  400.00
#> 30   0  0.00  381.90
#> 31   0  0.00  364.60
#> 32   0  0.00  348.10
#> 33   0  0.00  332.40
#> 34   0  0.00  317.30
#> 35   0  0.00  303.00
#> 36   0  0.00  289.30
#> 37   0  0.00  276.20
#> 38   0  0.00  263.70
#> 39   0  0.00  251.80
#> 40   0  0.00  240.40
#> 41   0  0.00  229.50
#> 42   0  0.00  219.10
#> 43   0  0.00  209.20
#> 44   0  0.00  199.80
#> 45   0  0.00  190.70
#> 46   0  0.00  182.10
#> 47   0  0.00  173.90
#> 48   0  0.00  166.00
#> 49   0  0.00  158.50
#> 50   0  0.00  151.30
#> 51   0  0.00  144.50
#> 52   0  0.00  137.90
#> 53   0  0.00  131.70
#> 54   0  0.00  125.70
#> 55   0  0.00  120.10
#> 56   0  0.00  114.60
#> 57   0  0.00  109.40
#> 58   0  0.00  104.50
#> 59   0  0.00   99.77
#> 60   0  0.00   95.26
#> 61   0  0.00   90.95
#> 62   0  0.00   86.83
#> 63   0  0.00   82.91
#> 64   0  0.00   79.16
#> 65   0  0.00   75.58
#> 66   0  0.00   72.16
#> 67   0  0.00   68.90
#> 68   0  0.00   65.78
#> 69   0  0.00   62.80
#> 70   0  0.00   59.96
#> 71   0  0.00   57.25
#> 72   0  0.00   54.66
#> 73   0  0.00   52.19
#> 74   0  0.00   49.83
#> 75   0  0.00   47.58
#> 76   0  0.00   45.42
#> 77   0  0.00   43.37
#> 78   0  0.00   41.41
#> 79   0  0.00   39.54
#> 80   0  0.00   37.75
#> 81   0  0.00   36.04
#> 82   0  0.00   34.41
#> 83   0  0.00   32.85
#> 84   0  0.00   31.37
#> 85   0  0.00   29.95
#> 86   0  0.00   28.59
#> 87   1  0.02   27.30
#> 88   1  0.16   26.07
#> 89   1  0.28   24.89
#> 90   1  0.40   23.76
#> 91   1  0.50   22.69
#> 92   1  0.59   21.66
#> 93   1  0.68   20.68
#> 94   2  0.76   19.75
#> 95   2  0.95   18.85
#> 96   2  1.12   18.00
#> 97   2  1.29   17.19
#> 98   2  1.43   16.41
#> 99   3  1.67   15.67
#> 100  3  1.91   14.96
#> 101  4  2.20   14.28
#> 102  3  2.47   13.64
#> 103  4  2.70   13.02
#> 104  4  2.95   12.43
#> 105  4  3.18   11.87
#> 106  4  3.39   11.33
#> 107  4  3.59   10.82
#> 108  4  3.77   10.33
#> 109  4  3.93    9.86
#> 110  4  4.08    9.42
#> 111  4  4.22    8.99
#> 112  4  4.35    8.58
#> 113  4  4.47    8.20
#> 114  5  4.60    7.82
#> 115  5  4.76    7.47
#> 116  5  4.91    7.13
#> 117  5  5.04    6.81
#> 118  5  5.16    6.50
#> 119  5  5.28    6.21
#> 120  6  5.39    5.93
#> 121  6  5.51    5.66
#> 122  6  5.62    5.40
#> 123  6  5.72    5.16
#> 124  6  5.81    4.93
#> 125  7  5.89    4.70
#> 126  7  6.00    4.49
#> 127  7  6.10    4.29
#> 128  7  6.19    4.09
#> 129  7  6.27    3.91
#> 130  7  6.35    3.73
#> 131  7  6.41    3.56
#> 132  8  6.48    3.40
#> 133  9  6.56    3.25
#> 134  9  6.64    3.10
#> 135  9  6.71    2.96
#> 136  9  6.78    2.83
#> 137 10  6.86    2.70
#> 138 11  7.07    2.58
#> 139 11  7.27    2.46
#> 140 12  7.44    2.35
#> 141 12  7.61    2.24
#> 142 12  7.76    2.14
#> 143 12  7.90    2.04
#> 144 12  8.03    1.95
#> 145 13  8.14    1.86
#> 146 13  8.24    1.78
#> 147 13  8.34    1.70
#> 148 13  8.43    1.62
#> 149 13  8.51    1.55
#> 150 14  8.58    1.48
#> 151 14  8.64    1.41
#> 152 14  8.71    1.35
#> 153 14  8.76    1.29
#> 154 14  8.81    1.23
#> 155 14  8.86    1.17
#> 156 14  8.90    1.12
#> 157 14  8.94    1.07
#> 158 14  8.98    1.02
#> 159 15  9.01    0.97
#> 160 16  9.10    0.93
#> 161 16  9.17    0.89
#> 162 16  9.24    0.85
#> 163 16  9.30    0.81
#> 164 16  9.37    0.77
#> 165 16  9.43    0.74
#> 166 17  9.48    0.71
#> 167 18  9.54    0.67
#> 168 19  9.59    0.64
#> 169 19  9.64    0.61
#> 170 19  9.69    0.59
#> 171 19  9.73    0.56
#> 172 20  9.78    0.53
#> 173 20  9.83    0.51
#> 174 20  9.87    0.49
#> 175 20  9.91    0.46
#> 176 20  9.95    0.44
#> 177 20  9.98    0.42
#> 178 20 10.02    0.40
#> 179 20 10.05    0.39
#> 180 20 10.08    0.37
#> 181 21 10.11    0.35
#> 182 21 10.14    0.34
#> 183 21 10.16    0.32
#> 184 21 10.19    0.31
#> 185 21 10.22    0.29
#> 186 21 10.24    0.28
#> 187 21 10.27    0.27
#> 188 21 10.29    0.25
#> 189 21 10.31    0.24
#> 190 21 10.33    0.23
#> 191 21 10.35    0.22
#> 192 21 10.36    0.21
#> 193 21 10.38    0.20
#> 194 21 10.39    0.19
#> 195 21 10.40    0.18
#> 196 21 10.42    0.18
#> 197 21 10.43    0.17
#> 198 21 10.44    0.16
#> 199 21 10.45    0.15
#> 200 21 10.46    0.15

max_t1$predictors

#> # A tibble: 1 x 6
#>   c1    c2      c3    c4    c5    f       
#>   <chr> <chr>   <chr> <chr> <chr> <chr>   
#> 1 aet   ppt_jja pH    awc   depth landform

max_t1$performance

#> # A tibble: 1 x 27
#>   regmult classes model threshold     thr_value n_presences n_absences TPR_mean
#>     <dbl> <fct>   <chr> <chr>             <dbl>       <int>      <int>    <dbl>
#> 1     0.6 lq      max   max_sens_spec     0.342         700        700    0.931
#> # ... with 19 more variables: TPR_sd <dbl>, TNR_mean <dbl>, TNR_sd <dbl>,
#> #   SORENSEN_mean <dbl>, SORENSEN_sd <dbl>, JACCARD_mean <dbl>,
#> #   JACCARD_sd <dbl>, FPB_mean <dbl>, FPB_sd <dbl>, OR_mean <dbl>, OR_sd <dbl>,
#> #   TSS_mean <dbl>, TSS_sd <dbl>, AUC_mean <dbl>, AUC_sd <dbl>,
#> #   BOYCE_mean <dbl>, BOYCE_sd <dbl>, IMAE_mean <dbl>, IMAE_sd <dbl>

max_t1$data_ens

#> # A tibble: 1,400 x 5
#>    rnames replicates part  pr_ab     pred
#>    <chr>  <chr>      <chr> <dbl>    <dbl>
#>  1 1      .part      1         0 0.628   
#>  2 3      .part      1         0 0.822   
#>  3 5      .part      1         0 0.727   
#>  4 7      .part      1         0 0.00287 
#>  5 8      .part      1         0 0.780   
#>  6 13     .part      1         0 0.150   
#>  7 15     .part      1         0 0.856   
#>  8 16     .part      1         0 0.000272
#>  9 22     .part      1         0 0.126   
#> 10 23     .part      1         0 0.395   
#> # ... with 1,390 more rows

# }